fix: 🐛 Improve multi-version events handling
Refactoring to avoid manage the same events on each contract version

BREAKING CHANGE: Y
