+++ src/contract_wrappers/modules/checkpoint/erc20_dividend_checkpoint_wrapper/3.0.0.ts
@@ -2,129 +2,22 @@ import {
   ERC20DividendCheckpointContract_3_0_0,
   ERC20DividendCheckpointEventArgs_3_0_0,
   ERC20DividendCheckpointEvents_3_0_0,
-  ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0,
-  ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0,
-  ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0,
-  ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0,
   LogWithDecodedArgs,
   Web3Wrapper,
 } from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
 import assert from '../../../../utils/assert';
-import { WithDividendCheckpoint_3_0_0 } from "../dividend_checkpoint_wrapper";
+import { WithDividendCheckpoint_3_0_0 } from '../dividend_checkpoint_wrapper';
 import ContractFactory from '../../../../factories/contractFactory';
-import {
-  GetLogsAsyncParams,
-  SubscribeAsyncParams,
-  EventCallback,
-  Subscribe,
-  GetLogs,
-  ContractVersion,
-  Constructor,
-} from '../../../../types';
-import ERC20DividendCheckpointCommon from './common';
+import { GetLogsAsyncParams, SubscribeAsyncParams, ContractVersion, Constructor } from '../../../../types';
+import ERC20DividendCheckpointCommon, {
+  ERC20DividendCheckpointSubscribeAsyncParams,
+  GetERC20DividendCheckpointLogsAsyncParams,
+} from './common';
 
-interface ERC20DividendDepositedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendDeposited;
-  callback: EventCallback<ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0>;
-}
-
-interface GetERC20DividendDepositedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendDeposited;
-}
-
-interface ERC20DividendClaimedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendClaimed;
-  callback: EventCallback<ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0>;
-}
-
-interface GetERC20DividendClaimedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendClaimed;
-}
-
-interface ERC20DividendReclaimedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendReclaimed;
-  callback: EventCallback<ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0>;
-}
-
-interface GetERC20DividendReclaimedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendReclaimed;
-}
-
-interface ERC20DividendWithholdingWithdrawnSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendWithholdingWithdrawn;
-  callback: EventCallback<ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0>;
-}
-
-interface GetERC20DividendWithholdingWithdrawnLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendWithholdingWithdrawn;
-}
-
-interface SetDefaultExcludedAddressesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
-  callback: EventCallback<ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>;
-}
-
-interface GetSetDefaultExcludedAddressesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
-}
-
-interface SetWithholdingSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholding;
-  callback: EventCallback<ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0>;
-}
-
-interface GetSetWithholdingLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholding;
-}
-
-interface SetWithholdingFixedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholdingFixed;
-  callback: EventCallback<ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0>;
-}
-
-interface GetSetWithholdingFixedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholdingFixed;
-}
-
-interface ERC20DividendCheckpointSubscribeAsyncParams extends Subscribe {
-  (params: ERC20DividendDepositedSubscribeAsyncParams): Promise<string>;
-  (params: ERC20DividendClaimedSubscribeAsyncParams): Promise<string>;
-  (params: ERC20DividendReclaimedSubscribeAsyncParams): Promise<string>;
-  (params: ERC20DividendWithholdingWithdrawnSubscribeAsyncParams): Promise<string>;
-  (params: SetDefaultExcludedAddressesSubscribeAsyncParams): Promise<string>;
-  (params: SetWithholdingSubscribeAsyncParams): Promise<string>;
-  (params: SetWithholdingFixedSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetERC20DividendCheckpointLogsAsyncParams extends GetLogs {
-  (params: GetERC20DividendDepositedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0>[]
-  >;
-  (params: GetERC20DividendClaimedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0>[]
-  >;
-  (params: GetERC20DividendReclaimedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0>[]
-  >;
-  (params: GetERC20DividendWithholdingWithdrawnLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0>[]
-  >;
-  (params: GetSetDefaultExcludedAddressesLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>[]
-  >;
-  (params: GetSetWithholdingLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0>[]
-  >;
-  (params: GetSetWithholdingFixedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0>[]
-  >;
-}
-
-const ERC20DividendCheckpointBase_3_0_0 = WithDividendCheckpoint_3_0_0(ERC20DividendCheckpointCommon as unknown as Constructor<ERC20DividendCheckpointCommon>);
+const ERC20DividendCheckpointBase_3_0_0 = WithDividendCheckpoint_3_0_0(
+  (ERC20DividendCheckpointCommon as unknown) as Constructor<ERC20DividendCheckpointCommon>,
+);
 
 export class ERC20DividendCheckpoint_3_0_0 extends ERC20DividendCheckpointBase_3_0_0 {
   public contract: Promise<ERC20DividendCheckpointContract_3_0_0>;
@@ -190,6 +83,8 @@ export class ERC20DividendCheckpoint_3_0_0 extends ERC20DividendCheckpointBase_3
   };
 }
 
-export function isERC20DividendCheckpoint_3_0_0(wrapper: ERC20DividendCheckpointCommon): wrapper is ERC20DividendCheckpoint_3_0_0 {
+export function isERC20DividendCheckpoint_3_0_0(
+  wrapper: ERC20DividendCheckpointCommon,
+): wrapper is ERC20DividendCheckpoint_3_0_0 {
   return wrapper.contractVersion === ContractVersion.V3_0_0;
-};
+}
