+++ src/contract_wrappers/modules/checkpoint/erc20_dividend_checkpoint_wrapper/common.ts
@@ -4,14 +4,32 @@ import {
   BigNumber,
   Web3Wrapper,
   PolyResponse,
+  ERC20DividendCheckpointEvents_3_0_0,
+  ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0,
+  ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0,
+  ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0,
+  ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0,
+  ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0,
+  ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0,
+  ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0,
+  ERC20DividendCheckpointPauseEventArgs_3_0_0,
+  ERC20DividendCheckpointSetWalletEventArgs_3_0_0,
+  ERC20DividendCheckpointUnpauseEventArgs_3_0_0,
+  ERC20DividendCheckpointUpdateDividendDatesEventArgs_3_0_0,
+  LogWithDecodedArgs,
 } from '@polymathnetwork/abi-wrappers';
 import assert from '../../../../utils/assert';
-import { DividendCheckpointCommon } from "../dividend_checkpoint_wrapper";
+import { DividendCheckpointCommon } from '../dividend_checkpoint_wrapper';
 import ContractFactory from '../../../../factories/contractFactory';
 import {
   TxParams,
   Perm,
   ErrorCode,
+  SubscribeAsyncParams,
+  EventCallback,
+  GetLogsAsyncParams,
+  Subscribe,
+  GetLogs,
 } from '../../../../types';
 import { numberToBigNumber, dateToBigNumber, stringToBytes32, valueToWei } from '../../../../utils/convert';
 import ContractWrapper from '../../../contract_wrapper';
@@ -21,6 +39,149 @@ export namespace ERC20DividendCheckpointTransactionParams {
   export interface CreateDividendWithCheckpoint extends CreateDividendWithCheckpointParams {}
   export interface CreateDividendWithExclusions extends CreateDividendWithExclusionsParams {}
   export interface CreateDividendWithCheckpointAndExclusions extends CreateDividendWithCheckpointAndExclusionsParams {}
+}
+
+interface ERC20DividendDepositedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendDeposited;
+  callback: EventCallback<ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0>;
+}
+
+interface GetERC20DividendDepositedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendDeposited;
+}
+
+interface ERC20DividendClaimedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendClaimed;
+  callback: EventCallback<ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0>;
+}
+
+interface GetERC20DividendClaimedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendClaimed;
+}
+
+interface ERC20DividendReclaimedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendReclaimed;
+  callback: EventCallback<ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0>;
+}
+
+interface GetERC20DividendReclaimedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendReclaimed;
+}
+
+interface ERC20DividendWithholdingWithdrawnSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendWithholdingWithdrawn;
+  callback: EventCallback<ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0>;
+}
+
+interface GetERC20DividendWithholdingWithdrawnLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.ERC20DividendWithholdingWithdrawn;
+}
+
+interface SetDefaultExcludedAddressesSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
+  callback: EventCallback<ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>;
+}
+
+interface GetSetDefaultExcludedAddressesLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
+}
+
+interface SetWithholdingSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholding;
+  callback: EventCallback<ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0>;
+}
+
+interface GetSetWithholdingLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholding;
+}
+
+interface SetWithholdingFixedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholdingFixed;
+  callback: EventCallback<ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0>;
+}
+
+interface GetSetWithholdingFixedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWithholdingFixed;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.Pause;
+  callback: EventCallback<ERC20DividendCheckpointPauseEventArgs_3_0_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.Unpause;
+  callback: EventCallback<ERC20DividendCheckpointUnpauseEventArgs_3_0_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.Unpause;
+}
+
+interface SetWalletSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWallet;
+  callback: EventCallback<ERC20DividendCheckpointSetWalletEventArgs_3_0_0>;
+}
+
+interface GetSetWalletLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.SetWallet;
+}
+
+interface UpdateDividendDatesSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.UpdateDividendDates;
+  callback: EventCallback<ERC20DividendCheckpointUpdateDividendDatesEventArgs_3_0_0>;
+}
+
+interface GetUpdateDividendDatesLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: ERC20DividendCheckpointEvents_3_0_0.UpdateDividendDates;
+}
+
+export interface ERC20DividendCheckpointSubscribeAsyncParams extends Subscribe {
+  (params: ERC20DividendDepositedSubscribeAsyncParams): Promise<string>;
+  (params: ERC20DividendClaimedSubscribeAsyncParams): Promise<string>;
+  (params: ERC20DividendReclaimedSubscribeAsyncParams): Promise<string>;
+  (params: ERC20DividendWithholdingWithdrawnSubscribeAsyncParams): Promise<string>;
+  (params: SetDefaultExcludedAddressesSubscribeAsyncParams): Promise<string>;
+  (params: SetWithholdingSubscribeAsyncParams): Promise<string>;
+  (params: SetWithholdingFixedSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+  (params: SetWalletSubscribeAsyncParams): Promise<string>;
+  (params: UpdateDividendDatesSubscribeAsyncParams): Promise<string>;
+}
+
+export interface GetERC20DividendCheckpointLogsAsyncParams extends GetLogs {
+  (params: GetERC20DividendDepositedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0>[]
+  >;
+  (params: GetERC20DividendClaimedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0>[]
+  >;
+  (params: GetERC20DividendReclaimedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0>[]
+  >;
+  (params: GetERC20DividendWithholdingWithdrawnLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0>[]
+  >;
+  (params: GetSetDefaultExcludedAddressesLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>[]
+  >;
+  (params: GetSetWithholdingLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0>[]
+  >;
+  (params: GetSetWithholdingFixedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<ERC20DividendCheckpointPauseEventArgs_3_0_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<ERC20DividendCheckpointUnpauseEventArgs_3_0_0>[]>;
+  (params: GetSetWalletLogsAsyncParams): Promise<LogWithDecodedArgs<ERC20DividendCheckpointSetWalletEventArgs_3_0_0>[]>;
+  (params: GetUpdateDividendDatesLogsAsyncParams): Promise<
+    LogWithDecodedArgs<ERC20DividendCheckpointUpdateDividendDatesEventArgs_3_0_0>[]
+  >;
 }
 
 /**
@@ -238,4 +399,4 @@ export default abstract class ERC20DividendCheckpointCommon extends DividendChec
 
 export function isERC20DividendCheckpoint(wrapper: ContractWrapper): wrapper is ERC20DividendCheckpointCommon {
   return wrapper instanceof ERC20DividendCheckpointCommon;
-};
+}
