+++ src/contract_wrappers/modules/checkpoint/ether_dividend_checkpoint_wrapper/3.0.0.ts
@@ -2,193 +2,22 @@ import {
   EtherDividendCheckpointContract_3_0_0,
   EtherDividendCheckpointEventArgs_3_0_0,
   EtherDividendCheckpointEvents_3_0_0,
-  EtherDividendCheckpointEtherDividendDepositedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendClaimedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendReclaimedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendClaimFailedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendWithholdingWithdrawnEventArgs_3_0_0,
-  EtherDividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0,
-  EtherDividendCheckpointSetWithholdingEventArgs_3_0_0,
-  EtherDividendCheckpointSetWithholdingFixedEventArgs_3_0_0,
-  EtherDividendCheckpointSetWalletEventArgs_3_0_0,
-  EtherDividendCheckpointUpdateDividendDatesEventArgs_3_0_0,
-  EtherDividendCheckpointPauseEventArgs_3_0_0,
-  EtherDividendCheckpointUnpauseEventArgs_3_0_0,
   Web3Wrapper,
   LogWithDecodedArgs,
 } from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
 import assert from '../../../../utils/assert';
-import { WithDividendCheckpoint_3_0_0 } from "../dividend_checkpoint_wrapper";
+import { WithDividendCheckpoint_3_0_0 } from '../dividend_checkpoint_wrapper';
 import ContractFactory from '../../../../factories/contractFactory';
-import {
-  GetLogsAsyncParams,
-  SubscribeAsyncParams,
-  EventCallback,
-  Subscribe,
-  GetLogs,
-  ContractVersion,
-  Constructor,
-} from '../../../../types';
-import EtherDividendCheckpointCommon from './common';
+import { GetLogsAsyncParams, SubscribeAsyncParams, ContractVersion, Constructor } from '../../../../types';
+import EtherDividendCheckpointCommon, {
+  EtherDividendCheckpointSubscribeAsyncParams,
+  GetEtherDividendCheckpointLogsAsyncParams,
+} from './common';
 
-interface EtherDividendDepositedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendDeposited;
-  callback: EventCallback<EtherDividendCheckpointEtherDividendDepositedEventArgs_3_0_0>;
-}
-
-interface GetEtherDividendDepositedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendDeposited;
-}
-
-interface EtherDividendClaimedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendClaimed;
-  callback: EventCallback<EtherDividendCheckpointEtherDividendClaimedEventArgs_3_0_0>;
-}
-
-interface GetEtherDividendClaimedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendClaimed;
-}
-
-interface EtherDividendReclaimedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendReclaimed;
-  callback: EventCallback<EtherDividendCheckpointEtherDividendReclaimedEventArgs_3_0_0>;
-}
-
-interface GetEtherDividendReclaimedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendReclaimed;
-}
-
-interface EtherDividendClaimFailedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendClaimFailed;
-  callback: EventCallback<EtherDividendCheckpointEtherDividendClaimFailedEventArgs_3_0_0>;
-}
-
-interface GetEtherDividendClaimFailedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendClaimFailed;
-}
-
-interface EtherDividendWithholdingWithdrawnSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendWithholdingWithdrawn;
-  callback: EventCallback<EtherDividendCheckpointEtherDividendWithholdingWithdrawnEventArgs_3_0_0>;
-}
-
-interface GetEtherDividendWithholdingWithdrawnLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.EtherDividendWithholdingWithdrawn;
-}
-
-interface SetDefaultExcludedAddressesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
-  callback: EventCallback<EtherDividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>;
-}
-
-interface GetSetDefaultExcludedAddressesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetDefaultExcludedAddresses;
-}
-
-interface SetWithholdingSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWithholding;
-  callback: EventCallback<EtherDividendCheckpointSetWithholdingEventArgs_3_0_0>;
-}
-
-interface GetSetWithholdingLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWithholding;
-}
-
-interface SetWithholdingFixedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWithholdingFixed;
-  callback: EventCallback<EtherDividendCheckpointSetWithholdingFixedEventArgs_3_0_0>;
-}
-
-interface GetSetWithholdingFixedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWithholdingFixed;
-}
-
-interface SetWalletSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWallet;
-  callback: EventCallback<EtherDividendCheckpointSetWalletEventArgs_3_0_0>;
-}
-
-interface GetSetWalletLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.SetWallet;
-}
-
-interface UpdateDividendDatesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.UpdateDividendDates;
-  callback: EventCallback<EtherDividendCheckpointUpdateDividendDatesEventArgs_3_0_0>;
-}
-
-interface GetUpdateDividendDatesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.UpdateDividendDates;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.Pause;
-  callback: EventCallback<EtherDividendCheckpointPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.Unpause;
-  callback: EventCallback<EtherDividendCheckpointUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: EtherDividendCheckpointEvents_3_0_0.Unpause;
-}
-
-interface EtherDividendCheckpointSubscribeAsyncParams extends Subscribe {
-  (params: EtherDividendDepositedSubscribeAsyncParams): Promise<string>;
-  (params: EtherDividendClaimedSubscribeAsyncParams): Promise<string>;
-  (params: EtherDividendReclaimedSubscribeAsyncParams): Promise<string>;
-  (params: EtherDividendClaimFailedSubscribeAsyncParams): Promise<string>;
-  (params: EtherDividendWithholdingWithdrawnSubscribeAsyncParams): Promise<string>;
-  (params: SetDefaultExcludedAddressesSubscribeAsyncParams): Promise<string>;
-  (params: SetWithholdingSubscribeAsyncParams): Promise<string>;
-  (params: SetWithholdingFixedSubscribeAsyncParams): Promise<string>;
-  (params: SetWalletSubscribeAsyncParams): Promise<string>;
-  (params: UpdateDividendDatesSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetEtherDividendCheckpointLogsAsyncParams extends GetLogs {
-  (params: GetEtherDividendDepositedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointEtherDividendDepositedEventArgs_3_0_0>[]
-  >;
-  (params: GetEtherDividendClaimedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointEtherDividendClaimedEventArgs_3_0_0>[]
-  >;
-  (params: GetEtherDividendReclaimedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointEtherDividendReclaimedEventArgs_3_0_0>[]
-  >;
-  (params: GetEtherDividendClaimFailedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointEtherDividendClaimFailedEventArgs_3_0_0>[]
-  >;
-  (params: GetEtherDividendWithholdingWithdrawnLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointEtherDividendWithholdingWithdrawnEventArgs_3_0_0>[]
-  >;
-  (params: GetSetDefaultExcludedAddressesLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0>[]
-  >;
-  (params: GetSetWithholdingLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointSetWithholdingEventArgs_3_0_0>[]
-  >;
-  (params: GetSetWithholdingFixedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointSetWithholdingFixedEventArgs_3_0_0>[]
-  >;
-  (params: GetSetWalletLogsAsyncParams): Promise<LogWithDecodedArgs<EtherDividendCheckpointSetWalletEventArgs_3_0_0>[]>;
-  (params: GetUpdateDividendDatesLogsAsyncParams): Promise<
-    LogWithDecodedArgs<EtherDividendCheckpointUpdateDividendDatesEventArgs_3_0_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<EtherDividendCheckpointPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<EtherDividendCheckpointUnpauseEventArgs_3_0_0>[]>;
-}
-
-const EtherDividendCheckpointBase_3_0_0 = WithDividendCheckpoint_3_0_0(EtherDividendCheckpointCommon as unknown as Constructor<EtherDividendCheckpointCommon>);
+const EtherDividendCheckpointBase_3_0_0 = WithDividendCheckpoint_3_0_0(
+  (EtherDividendCheckpointCommon as unknown) as Constructor<EtherDividendCheckpointCommon>,
+);
 
 export class EtherDividendCheckpoint_3_0_0 extends EtherDividendCheckpointBase_3_0_0 {
   public contract: Promise<EtherDividendCheckpointContract_3_0_0>;
@@ -254,6 +83,8 @@ export class EtherDividendCheckpoint_3_0_0 extends EtherDividendCheckpointBase_3
   };
 }
 
-export function isEtherDividendCheckpoint_3_0_0(wrapper: EtherDividendCheckpointCommon): wrapper is EtherDividendCheckpoint_3_0_0 {
+export function isEtherDividendCheckpoint_3_0_0(
+  wrapper: EtherDividendCheckpointCommon,
+): wrapper is EtherDividendCheckpoint_3_0_0 {
   return wrapper.contractVersion === ContractVersion.V3_0_0;
-};
+}
