+++ src/contract_wrappers/modules/sto/capped_sto_wrapper/3.0.0.ts
@@ -1,76 +1,24 @@
-import { CappedSTOContract_3_0_0, Web3Wrapper, CappedSTOEventArgs_3_0_0, CappedSTOEvents_3_0_0, LogWithDecodedArgs, CappedSTOTokenPurchaseEventArgs_3_0_0, CappedSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0, CappedSTOSetFundRaiseTypesEventArgs_3_0_0, CappedSTOPauseEventArgs_3_0_0, CappedSTOUnpauseEventArgs_3_0_0, BigNumber } from '@polymathnetwork/abi-wrappers';
+import {
+  CappedSTOContract_3_0_0,
+  CappedSTOEventArgs_3_0_0,
+  CappedSTOEvents_3_0_0,
+  Web3Wrapper,
+  LogWithDecodedArgs,
+  BigNumber,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import CappedSTOCommon from './common';
+import CappedSTOCommon, { GetCappedSTOLogsAsyncParams, CappedSTOSubscribeAsyncParams } from './common';
 import assert from '../../../../utils/assert';
 import {
-  ContractVersion, SubscribeAsyncParams, GetLogsAsyncParams, Subscribe, GetLogs, FULL_DECIMALS, Constructor, EventCallback
+  ContractVersion,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  FULL_DECIMALS,
+  Constructor,
 } from '../../../../types';
 import ContractFactory from '../../../../factories/contractFactory';
 import { WithSTO_3_0_0 } from '../sto_wrapper';
 import { bigNumberToDate, weiToValue } from '../../../../utils/convert';
-
-interface TokenPurchaseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.TokenPurchase;
-  callback: EventCallback<CappedSTOTokenPurchaseEventArgs_3_0_0>;
-}
-
-interface GetTokenPurchaseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.TokenPurchase;
-}
-
-interface SetAllowBeneficialInvestmentsSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.SetAllowBeneficialInvestments;
-  callback: EventCallback<CappedSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0>;
-}
-
-interface GetSetAllowBeneficialInvestmentsLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.SetAllowBeneficialInvestments;
-}
-
-interface SetFundRaiseTypesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.SetFundRaiseTypes;
-  callback: EventCallback<CappedSTOSetFundRaiseTypesEventArgs_3_0_0>;
-}
-
-interface GetSetFundRaiseTypesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.SetFundRaiseTypes;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.Pause;
-  callback: EventCallback<CappedSTOPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.Unpause;
-  callback: EventCallback<CappedSTOUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CappedSTOEvents_3_0_0.Unpause;
-}
-
-interface CappedSTOSubscribeAsyncParams extends Subscribe {
-  (params: TokenPurchaseSubscribeAsyncParams): Promise<string>;
-  (params: SetAllowBeneficialInvestmentsSubscribeAsyncParams): Promise<string>;
-  (params: SetFundRaiseTypesSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetCappedSTOLogsAsyncParams extends GetLogs {
-  (params: GetTokenPurchaseLogsAsyncParams): Promise<LogWithDecodedArgs<CappedSTOTokenPurchaseEventArgs_3_0_0>[]>;
-  (params: GetSetAllowBeneficialInvestmentsLogsAsyncParams): Promise<
-    LogWithDecodedArgs<CappedSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0>[]
-  >;
-  (params: GetSetFundRaiseTypesLogsAsyncParams): Promise<LogWithDecodedArgs<CappedSTOSetFundRaiseTypesEventArgs_3_0_0>[]>;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<CappedSTOPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<CappedSTOUnpauseEventArgs_3_0_0>[]>;
-}
 
 // // Return types ////
 export interface CappedSTODetails {
@@ -93,7 +41,7 @@ export interface CappedSTODetails {
 }
 // // End of return types ////
 
-const CappedSTOBase_3_0_0 = WithSTO_3_0_0(CappedSTOCommon as unknown as Constructor<CappedSTOCommon>);
+const CappedSTOBase_3_0_0 = WithSTO_3_0_0((CappedSTOCommon as unknown) as Constructor<CappedSTOCommon>);
 
 export class CappedSTO_3_0_0 extends CappedSTOBase_3_0_0 {
   public contract: Promise<CappedSTOContract_3_0_0>;
@@ -105,7 +53,11 @@ export class CappedSTO_3_0_0 extends CappedSTOBase_3_0_0 {
    * @param web3Wrapper Web3Wrapper instance to use
    * @param contract
    */
-  public constructor(web3Wrapper: Web3Wrapper, contract: Promise<CappedSTOContract_3_0_0>, contractFactory: ContractFactory) {
+  public constructor(
+    web3Wrapper: Web3Wrapper,
+    contract: Promise<CappedSTOContract_3_0_0>,
+    contractFactory: ContractFactory,
+  ) {
     super(web3Wrapper, contract, contractFactory);
     this.contract = contract;
   }
@@ -175,4 +127,4 @@ export class CappedSTO_3_0_0 extends CappedSTOBase_3_0_0 {
 
 export function isCappedSTO_3_0_0(wrapper: CappedSTOCommon): wrapper is CappedSTO_3_0_0 {
   return wrapper.contractVersion === ContractVersion.V3_0_0;
-};
+}
