+++ src/contract_wrappers/modules/sto/usd_tiered_sto_wrapper/3.0.0.ts
@@ -1,169 +1,30 @@
-import { USDTieredSTOContract_3_0_0, BigNumber, Web3Wrapper, USDTieredSTOEvents_3_0_0, USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0, USDTieredSTOSetNonAccreditedLimitEventArgs_3_0_0, USDTieredSTOSetTreasuryWalletEventArgs_3_0_0, USDTieredSTOTokenPurchaseEventArgs_3_0_0, USDTieredSTOFundsReceivedEventArgs_3_0_0, USDTieredSTOReserveTokenMintEventArgs_3_0_0, USDTieredSTOSetAddressesEventArgs_3_0_0, USDTieredSTOSetLimitsEventArgs_3_0_0, USDTieredSTOSetTimesEventArgs_3_0_0, USDTieredSTOSetTiersEventArgs_3_0_0, USDTieredSTOSetFundRaiseTypesEventArgs_3_0_0, USDTieredSTOPauseEventArgs_3_0_0, USDTieredSTOUnpauseEventArgs_3_0_0, LogWithDecodedArgs, USDTieredSTOEventArgs_3_0_0 } from '@polymathnetwork/abi-wrappers';
+import {
+  USDTieredSTOContract_3_0_0,
+  USDTieredSTOEvents_3_0_0,
+  USDTieredSTOEventArgs_3_0_0,
+  BigNumber,
+  Web3Wrapper,
+  LogWithDecodedArgs,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import USDTieredSTOCommon, { TierIndexParams } from './common';
+import USDTieredSTOCommon, {
+  TierIndexParams,
+  USDTieredSTOSubscribeAsyncParams,
+  GetUSDTieredSTOLogsAsyncParams,
+} from './common';
 import assert from '../../../../utils/assert';
 import {
-  ErrorCode, ContractVersion, SubscribeAsyncParams, GetLogsAsyncParams, Subscribe, GetLogs, TxParams, FULL_DECIMALS, Constructor, EventCallback
+  ErrorCode,
+  ContractVersion,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  TxParams,
+  FULL_DECIMALS,
+  Constructor,
 } from '../../../../types';
-import {
-  numberToBigNumber,  
-  weiToValue,
-} from '../../../../utils/convert';
+import { numberToBigNumber, weiToValue } from '../../../../utils/convert';
 import ContractFactory from '../../../../factories/contractFactory';
 import { WithSTO_3_0_0 } from '../sto_wrapper';
-
-interface SetAllowBeneficialInvestmentsSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetAllowBeneficialInvestments;
-  callback: EventCallback<USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0>;
-}
-
-interface GetSetAllowBeneficialInvestmentsLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetAllowBeneficialInvestments;
-}
-
-interface SetNonAccreditedLimitSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetNonAccreditedLimit;
-  callback: EventCallback<USDTieredSTOSetNonAccreditedLimitEventArgs_3_0_0>;
-}
-
-interface GetSetNonAccreditedLimitLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetNonAccreditedLimit;
-}
-
-interface SetTreasuryWalletSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTreasuryWallet;
-  callback: EventCallback<USDTieredSTOSetTreasuryWalletEventArgs_3_0_0>;
-}
-
-interface GetSetTreasuryWalletLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTreasuryWallet;
-}
-
-interface TokenPurchaseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.TokenPurchase;
-  callback: EventCallback<USDTieredSTOTokenPurchaseEventArgs_3_0_0>;
-}
-
-interface GetTokenPurchaseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.TokenPurchase;
-}
-
-interface FundsReceivedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.FundsReceived;
-  callback: EventCallback<USDTieredSTOFundsReceivedEventArgs_3_0_0>;
-}
-
-interface GetFundsReceivedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.FundsReceived;
-}
-
-interface ReserveTokenMintSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.ReserveTokenMint;
-  callback: EventCallback<USDTieredSTOReserveTokenMintEventArgs_3_0_0>;
-}
-
-interface GetReserveTokenMintLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.ReserveTokenMint;
-}
-
-interface SetAddressesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetAddresses;
-  callback: EventCallback<USDTieredSTOSetAddressesEventArgs_3_0_0>;
-}
-
-interface GetSetAddressesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetAddresses;
-}
-
-interface SetLimitsSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetLimits;
-  callback: EventCallback<USDTieredSTOSetLimitsEventArgs_3_0_0>;
-}
-
-interface GetSetLimitsLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetLimits;
-}
-
-interface SetTimesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTimes;
-  callback: EventCallback<USDTieredSTOSetTimesEventArgs_3_0_0>;
-}
-
-interface GetSetTimesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTimes;
-}
-
-interface SetTiersSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTiers;
-  callback: EventCallback<USDTieredSTOSetTiersEventArgs_3_0_0>;
-}
-
-interface GetSetTiersLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetTiers;
-}
-
-interface SetFundRaiseTypesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetFundRaiseTypes;
-  callback: EventCallback<USDTieredSTOSetFundRaiseTypesEventArgs_3_0_0>;
-}
-
-interface GetSetFundRaiseTypesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.SetFundRaiseTypes;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.Pause;
-  callback: EventCallback<USDTieredSTOPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.Unpause;
-  callback: EventCallback<USDTieredSTOUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_0_0.Unpause;
-}
-
-interface USDTieredSTOSubscribeAsyncParams extends Subscribe {
-  (params: SetAllowBeneficialInvestmentsSubscribeAsyncParams): Promise<string>;
-  (params: SetNonAccreditedLimitSubscribeAsyncParams): Promise<string>;
-  (params: SetTreasuryWalletSubscribeAsyncParams): Promise<string>;
-  (params: TokenPurchaseSubscribeAsyncParams): Promise<string>;
-  (params: FundsReceivedSubscribeAsyncParams): Promise<string>;
-  (params: ReserveTokenMintSubscribeAsyncParams): Promise<string>;
-  (params: SetAddressesSubscribeAsyncParams): Promise<string>;
-  (params: SetLimitsSubscribeAsyncParams): Promise<string>;
-  (params: SetTimesSubscribeAsyncParams): Promise<string>;
-  (params: SetTiersSubscribeAsyncParams): Promise<string>;
-  (params: SetFundRaiseTypesSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetUSDTieredSTOLogsAsyncParams extends GetLogs {
-  (params: GetSetAllowBeneficialInvestmentsLogsAsyncParams): Promise<
-    LogWithDecodedArgs<USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0>[]
-  >;
-  (params: GetSetNonAccreditedLimitLogsAsyncParams): Promise<
-    LogWithDecodedArgs<USDTieredSTOSetNonAccreditedLimitEventArgs_3_0_0>[]
-  >;
-  (params: GetSetTreasuryWalletLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTreasuryWalletEventArgs_3_0_0>[]>;
-  (params: GetTokenPurchaseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOTokenPurchaseEventArgs_3_0_0>[]>;
-  (params: GetFundsReceivedLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOFundsReceivedEventArgs_3_0_0>[]>;
-  (params: GetReserveTokenMintLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOReserveTokenMintEventArgs_3_0_0>[]>;
-  (params: GetSetAddressesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetAddressesEventArgs_3_0_0>[]>;
-  (params: GetSetLimitsLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetLimitsEventArgs_3_0_0>[]>;
-  (params: GetSetTimesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTimesEventArgs_3_0_0>[]>;
-  (params: GetSetTiersLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTiersEventArgs_3_0_0>[]>;
-  (params: GetSetFundRaiseTypesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetFundRaiseTypesEventArgs_3_0_0>[]>;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOUnpauseEventArgs_3_0_0>[]>;
-}
 
 interface MintedByTier {
   mintedInETH: BigNumber;
@@ -186,7 +47,7 @@ interface Tier {
   mintedDiscountPoly: BigNumber;
 }
 
-const USDTieredSTOBase_3_0_0 = WithSTO_3_0_0(USDTieredSTOCommon as unknown as Constructor<USDTieredSTOCommon>);
+const USDTieredSTOBase_3_0_0 = WithSTO_3_0_0((USDTieredSTOCommon as unknown) as Constructor<USDTieredSTOCommon>);
 
 export class USDTieredSTO_3_0_0 extends USDTieredSTOBase_3_0_0 {
   public contract: Promise<USDTieredSTOContract_3_0_0>;
@@ -198,7 +59,11 @@ export class USDTieredSTO_3_0_0 extends USDTieredSTOBase_3_0_0 {
    * @param web3Wrapper Web3Wrapper instance to use
    * @param contract
    */
-  public constructor(web3Wrapper: Web3Wrapper, contract: Promise<USDTieredSTOContract_3_0_0>, contractFactory: ContractFactory) {
+  public constructor(
+    web3Wrapper: Web3Wrapper,
+    contract: Promise<USDTieredSTOContract_3_0_0>,
+    contractFactory: ContractFactory,
+  ) {
     super(web3Wrapper, contract, contractFactory);
     this.contract = contract;
   }
@@ -248,7 +113,7 @@ export class USDTieredSTO_3_0_0 extends USDTieredSTOBase_3_0_0 {
       mintedDiscountPoly: weiToValue(result[5], decimals),
     };
     return typedResult;
-  }; 
+  };
 
   /**
    * Return array of minted tokens in each fund raise type for given tier
@@ -323,4 +188,4 @@ export class USDTieredSTO_3_0_0 extends USDTieredSTOBase_3_0_0 {
 
 export function isUSDTieredSTO_3_0_0(wrapper: USDTieredSTOCommon): wrapper is USDTieredSTO_3_0_0 {
   return wrapper.contractVersion === ContractVersion.V3_0_0;
-};
+}
