+++ src/contract_wrappers/modules/sto/usd_tiered_sto_wrapper/3.1.0.ts
@@ -1,135 +1,33 @@
-import { USDTieredSTOContract_3_1_0, BigNumber, Web3Wrapper, USDTieredSTOEvents_3_1_0, USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_1_0, USDTieredSTOSetNonAccreditedLimitEventArgs_3_1_0, USDTieredSTOSetTreasuryWalletEventArgs_3_1_0, USDTieredSTOTokenPurchaseEventArgs_3_1_0, USDTieredSTOFundsReceivedEventArgs_3_1_0, USDTieredSTOReserveTokenMintEventArgs_3_1_0, USDTieredSTOSetAddressesEventArgs_3_1_0, USDTieredSTOSetLimitsEventArgs_3_1_0, USDTieredSTOSetTimesEventArgs_3_1_0, USDTieredSTOSetTiersEventArgs_3_1_0, USDTieredSTOSetFundRaiseTypesEventArgs_3_1_0, USDTieredSTOPauseEventArgs_3_1_0, USDTieredSTOUnpauseEventArgs_3_1_0, LogWithDecodedArgs, USDTieredSTOEventArgs_3_1_0, USDTieredSTOReserveTokenTransferEventArgs_3_1_0, USDTieredSTOAllowPreMintFlagEventArgs_3_1_0, USDTieredSTORevokePreMintFlagEventArgs_3_1_0 } from '@polymathnetwork/abi-wrappers';
+import {
+  USDTieredSTOContract_3_1_0,
+  USDTieredSTOEvents_3_1_0,
+  USDTieredSTOEventArgs_3_1_0,
+  USDTieredSTOReserveTokenTransferEventArgs_3_1_0,
+  USDTieredSTOAllowPreMintFlagEventArgs_3_1_0,
+  USDTieredSTORevokePreMintFlagEventArgs_3_1_0,
+  LogWithDecodedArgs,
+  BigNumber,
+  Web3Wrapper,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import USDTieredSTOCommon, { TierIndexParams } from './common';
+import USDTieredSTOCommon, {
+  TierIndexParams,
+  USDTieredSTOSubscribeAsyncParams,
+  GetUSDTieredSTOLogsAsyncParams,
+} from './common';
 import assert from '../../../../utils/assert';
 import {
-  ErrorCode, ContractVersion, SubscribeAsyncParams, GetLogsAsyncParams, Subscribe, GetLogs, FULL_DECIMALS, Constructor, EventCallback
+  ErrorCode,
+  ContractVersion,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  FULL_DECIMALS,
+  Constructor,
+  EventCallback,
 } from '../../../../types';
-import {
-  numberToBigNumber,  
-  weiToValue,
-  bigNumberToDate,
-  weiArrayToValueArray,
-} from '../../../../utils/convert';
+import { numberToBigNumber, weiToValue, bigNumberToDate, weiArrayToValueArray } from '../../../../utils/convert';
 import ContractFactory from '../../../../factories/contractFactory';
 import { WithSTO_3_1_0 } from '../sto_wrapper';
-
-interface SetAllowBeneficialInvestmentsSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetAllowBeneficialInvestments;
-  callback: EventCallback<USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_1_0>;
-}
-
-interface GetSetAllowBeneficialInvestmentsLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetAllowBeneficialInvestments;
-}
-
-interface SetNonAccreditedLimitSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetNonAccreditedLimit;
-  callback: EventCallback<USDTieredSTOSetNonAccreditedLimitEventArgs_3_1_0>;
-}
-
-interface GetSetNonAccreditedLimitLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetNonAccreditedLimit;
-}
-
-interface SetTreasuryWalletSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTreasuryWallet;
-  callback: EventCallback<USDTieredSTOSetTreasuryWalletEventArgs_3_1_0>;
-}
-
-interface GetSetTreasuryWalletLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTreasuryWallet;
-}
-
-interface TokenPurchaseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.TokenPurchase;
-  callback: EventCallback<USDTieredSTOTokenPurchaseEventArgs_3_1_0>;
-}
-
-interface GetTokenPurchaseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.TokenPurchase;
-}
-
-interface FundsReceivedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.FundsReceived;
-  callback: EventCallback<USDTieredSTOFundsReceivedEventArgs_3_1_0>;
-}
-
-interface GetFundsReceivedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.FundsReceived;
-}
-
-interface ReserveTokenMintSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.ReserveTokenMint;
-  callback: EventCallback<USDTieredSTOReserveTokenMintEventArgs_3_1_0>;
-}
-
-interface GetReserveTokenMintLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.ReserveTokenMint;
-}
-
-interface SetAddressesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetAddresses;
-  callback: EventCallback<USDTieredSTOSetAddressesEventArgs_3_1_0>;
-}
-
-interface GetSetAddressesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetAddresses;
-}
-
-interface SetLimitsSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetLimits;
-  callback: EventCallback<USDTieredSTOSetLimitsEventArgs_3_1_0>;
-}
-
-interface GetSetLimitsLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetLimits;
-}
-
-interface SetTimesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTimes;
-  callback: EventCallback<USDTieredSTOSetTimesEventArgs_3_1_0>;
-}
-
-interface GetSetTimesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTimes;
-}
-
-interface SetTiersSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTiers;
-  callback: EventCallback<USDTieredSTOSetTiersEventArgs_3_1_0>;
-}
-
-interface GetSetTiersLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetTiers;
-}
-
-interface SetFundRaiseTypesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetFundRaiseTypes;
-  callback: EventCallback<USDTieredSTOSetFundRaiseTypesEventArgs_3_1_0>;
-}
-
-interface GetSetFundRaiseTypesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.SetFundRaiseTypes;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.Pause;
-  callback: EventCallback<USDTieredSTOPauseEventArgs_3_1_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.Unpause;
-  callback: EventCallback<USDTieredSTOUnpauseEventArgs_3_1_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: USDTieredSTOEvents_3_1_0.Unpause;
-}
 
 interface ReserveTokenTransferSubscribeAsyncParams extends SubscribeAsyncParams {
   eventName: USDTieredSTOEvents_3_1_0.ReserveTokenTransfer;
@@ -158,46 +56,22 @@ interface GetRevokePreMintFlagLogsAsyncParams extends GetLogsAsyncParams {
   eventName: USDTieredSTOEvents_3_1_0.RevokePreMintFlag;
 }
 
-interface USDTieredSTOSubscribeAsyncParams extends Subscribe {
-  (params: SetAllowBeneficialInvestmentsSubscribeAsyncParams): Promise<string>;
-  (params: SetNonAccreditedLimitSubscribeAsyncParams): Promise<string>;
-  (params: SetTreasuryWalletSubscribeAsyncParams): Promise<string>;
-  (params: TokenPurchaseSubscribeAsyncParams): Promise<string>;
-  (params: FundsReceivedSubscribeAsyncParams): Promise<string>;
-  (params: ReserveTokenMintSubscribeAsyncParams): Promise<string>;
-  (params: SetAddressesSubscribeAsyncParams): Promise<string>;
-  (params: SetLimitsSubscribeAsyncParams): Promise<string>;
-  (params: SetTimesSubscribeAsyncParams): Promise<string>;
-  (params: SetTiersSubscribeAsyncParams): Promise<string>;
-  (params: SetFundRaiseTypesSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+interface USDTieredSTOSubscribeAsyncParams_3_1_0 extends USDTieredSTOSubscribeAsyncParams {
   (params: ReserveTokenTransferSubscribeAsyncParams): Promise<string>;
   (params: AllowPreMintFlagSubscribeAsyncParams): Promise<string>;
   (params: RevokePreMintFlagSubscribeAsyncParams): Promise<string>;
 }
 
-interface GetUSDTieredSTOLogsAsyncParams extends GetLogs {
-  (params: GetSetAllowBeneficialInvestmentsLogsAsyncParams): Promise<
-    LogWithDecodedArgs<USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_1_0>[]
+interface GetUSDTieredSTOLogsAsyncParams_3_1_0 extends GetUSDTieredSTOLogsAsyncParams {
+  (params: GetReserveTokenTransferLogsAsyncParams): Promise<
+    LogWithDecodedArgs<USDTieredSTOReserveTokenTransferEventArgs_3_1_0>[]
   >;
-  (params: GetSetNonAccreditedLimitLogsAsyncParams): Promise<
-    LogWithDecodedArgs<USDTieredSTOSetNonAccreditedLimitEventArgs_3_1_0>[]
+  (params: GetAllowPreMintFlagLogsAsyncParams): Promise<
+    LogWithDecodedArgs<USDTieredSTOAllowPreMintFlagEventArgs_3_1_0>[]
   >;
-  (params: GetSetTreasuryWalletLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTreasuryWalletEventArgs_3_1_0>[]>;
-  (params: GetTokenPurchaseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOTokenPurchaseEventArgs_3_1_0>[]>;
-  (params: GetFundsReceivedLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOFundsReceivedEventArgs_3_1_0>[]>;
-  (params: GetReserveTokenMintLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOReserveTokenMintEventArgs_3_1_0>[]>;
-  (params: GetSetAddressesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetAddressesEventArgs_3_1_0>[]>;
-  (params: GetSetLimitsLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetLimitsEventArgs_3_1_0>[]>;
-  (params: GetSetTimesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTimesEventArgs_3_1_0>[]>;
-  (params: GetSetTiersLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetTiersEventArgs_3_1_0>[]>;
-  (params: GetSetFundRaiseTypesLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOSetFundRaiseTypesEventArgs_3_1_0>[]>;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOPauseEventArgs_3_1_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOUnpauseEventArgs_3_1_0>[]>;
-  (params: GetReserveTokenTransferLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOReserveTokenTransferEventArgs_3_1_0>[]>;
-  (params: GetAllowPreMintFlagLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTOAllowPreMintFlagEventArgs_3_1_0>[]>;
-  (params: GetRevokePreMintFlagLogsAsyncParams): Promise<LogWithDecodedArgs<USDTieredSTORevokePreMintFlagEventArgs_3_1_0>[]>;
+  (params: GetRevokePreMintFlagLogsAsyncParams): Promise<
+    LogWithDecodedArgs<USDTieredSTORevokePreMintFlagEventArgs_3_1_0>[]
+  >;
 }
 
 interface MintedByTier {
@@ -248,7 +122,7 @@ interface Tier {
   soldDiscountPoly: BigNumber;
 }
 
-const USDTieredSTOBase_3_1_0 = WithSTO_3_1_0(USDTieredSTOCommon as unknown as Constructor<USDTieredSTOCommon>);
+const USDTieredSTOBase_3_1_0 = WithSTO_3_1_0((USDTieredSTOCommon as unknown) as Constructor<USDTieredSTOCommon>);
 
 export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
   public contract: Promise<USDTieredSTOContract_3_1_0>;
@@ -260,7 +134,11 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
    * @param web3Wrapper Web3Wrapper instance to use
    * @param contract
    */
-  public constructor(web3Wrapper: Web3Wrapper, contract: Promise<USDTieredSTOContract_3_1_0>, contractFactory: ContractFactory) {
+  public constructor(
+    web3Wrapper: Web3Wrapper,
+    contract: Promise<USDTieredSTOContract_3_1_0>,
+    contractFactory: ContractFactory,
+  ) {
     super(web3Wrapper, contract, contractFactory);
     this.contract = contract;
   }
@@ -308,7 +186,7 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
       soldDiscountPoly: weiToValue(result[5], decimals),
     };
     return typedResult;
-  };  
+  };
 
   /**
    * Return the total number of tokens sold in a given tier
@@ -321,7 +199,7 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
       await (await this.contract).getTotalTokensSoldByTier.callAsync(numberToBigNumber(params.tier)),
       await (await this.securityTokenContract()).decimals.callAsync(),
     );
-  };  
+  };
 
   /**
    * Return array of sold tokens in each fund raise type for given tier
@@ -343,7 +221,7 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
    * Subscribe to an event type emitted by the contract.
    * @return Subscription token used later to unsubscribe
    */
-  public subscribeAsync: USDTieredSTOSubscribeAsyncParams = async <ArgsType extends USDTieredSTOEventArgs_3_1_0>(
+  public subscribeAsync: USDTieredSTOSubscribeAsyncParams_3_1_0 = async <ArgsType extends USDTieredSTOEventArgs_3_1_0>(
     params: SubscribeAsyncParams,
   ): Promise<string> => {
     assert.doesBelongToStringEnum('eventName', params.eventName, USDTieredSTOEvents_3_1_0);
@@ -364,7 +242,7 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
    * Gets historical logs without creating a subscription
    * @return Array of logs that match the parameters
    */
-  public getLogsAsync: GetUSDTieredSTOLogsAsyncParams = async <ArgsType extends USDTieredSTOEventArgs_3_1_0>(
+  public getLogsAsync: GetUSDTieredSTOLogsAsyncParams_3_1_0 = async <ArgsType extends USDTieredSTOEventArgs_3_1_0>(
     params: GetLogsAsyncParams,
   ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
     assert.doesBelongToStringEnum('eventName', params.eventName, USDTieredSTOEvents_3_1_0);
@@ -381,4 +259,4 @@ export class USDTieredSTO_3_1_0 extends USDTieredSTOBase_3_1_0 {
 
 export function isUSDTieredSTO_3_1_0(wrapper: USDTieredSTOCommon): wrapper is USDTieredSTO_3_1_0 {
   return wrapper.contractVersion === ContractVersion.V3_1_0;
-};
+}
