+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/3.0.0.ts
@@ -1,52 +1,20 @@
-import { CountTransferManagerEvents_3_0_0, CountTransferManagerModifyHolderCountEventArgs_3_0_0, CountTransferManagerPauseEventArgs_3_0_0, CountTransferManagerUnpauseEventArgs_3_0_0, LogWithDecodedArgs, CountTransferManagerEventArgs_3_0_0 } from '@polymathnetwork/abi-wrappers';
+import {
+  CountTransferManagerEvents_3_0_0,
+  LogWithDecodedArgs,
+  CountTransferManagerEventArgs_3_0_0,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import { ContractVersion, SubscribeAsyncParams, EventCallback, GetLogsAsyncParams, Subscribe, GetLogs, Constructor } from '../../../../types';
-import CountTransferManagerCommon from './common';
+import { ContractVersion, SubscribeAsyncParams, GetLogsAsyncParams, Constructor } from '../../../../types';
+import CountTransferManagerCommon, {
+  CountTransferManagerSubscribeAsyncParams,
+  GetCountTransferManagerLogsAsyncParams,
+} from './common';
 import assert from '../../../../utils/assert';
 import { WithModule_3_0_0 } from '../../module_wrapper';
 
-interface ModifyHolderCountSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
-  callback: EventCallback<CountTransferManagerModifyHolderCountEventArgs_3_0_0>;
-}
-
-interface GetModifyHolderCountLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Pause;
-  callback: EventCallback<CountTransferManagerPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Unpause;
-  callback: EventCallback<CountTransferManagerUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Unpause;
-}
-
-interface CountTransferManagerSubscribeAsyncParams extends Subscribe {
-  (params: ModifyHolderCountSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetCountTransferManagerLogsAsyncParams extends GetLogs {
-  (params: GetModifyHolderCountLogsAsyncParams): Promise<
-    LogWithDecodedArgs<CountTransferManagerModifyHolderCountEventArgs_3_0_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerUnpauseEventArgs_3_0_0>[]>;
-}
-
-const CountTransferManagerBase_3_0_0 = WithModule_3_0_0(CountTransferManagerCommon as unknown as Constructor<CountTransferManagerCommon>);
+const CountTransferManagerBase_3_0_0 = WithModule_3_0_0((CountTransferManagerCommon as unknown) as Constructor<
+  CountTransferManagerCommon
+>);
 
 export class CountTransferManager_3_0_0 extends CountTransferManagerBase_3_0_0 {
   public contractVersion = ContractVersion.V3_0_0;
