+++ src/contract_wrappers/modules/transfer_manager/lock_up_transfer_manager_wrapper/common.ts
@@ -2,7 +2,16 @@ import {
   BigNumber,
   LockUpTransferManagerContract_3_0_0,
   Web3Wrapper,
-  PolyResponse
+  PolyResponse,
+  LogWithDecodedArgs,
+  LockUpTransferManagerEvents_3_0_0,
+  LockUpTransferManagerAddLockUpToUserEventArgs_3_0_0,
+  LockUpTransferManagerRemoveLockUpFromUserEventArgs_3_0_0,
+  LockUpTransferManagerModifyLockUpTypeEventArgs_3_0_0,
+  LockUpTransferManagerAddNewLockUpTypeEventArgs_3_0_0,
+  LockUpTransferManagerRemoveLockUpTypeEventArgs_3_0_0,
+  LockUpTransferManagerPauseEventArgs_3_0_0,
+  LockUpTransferManagerUnpauseEventArgs_3_0_0,
 } from '@polymathnetwork/abi-wrappers';
 import assert from '../../../../utils/assert';
 import { ModuleCommon } from '../../module_wrapper';
@@ -13,6 +22,11 @@ import {
   TransferResult,
   TxParams,
   ErrorCode,
+  GetLogs,
+  Subscribe,
+  GetLogsAsyncParams,
+  SubscribeAsyncParams,
+  EventCallback,
 } from '../../../../types';
 import {
   bigNumberToDate,
@@ -44,6 +58,99 @@ export namespace LockUpTransferManagerTransactionParams {
   export interface RemoveLockUpFromUserMulti extends RemoveLockUpFromUserMultiParams {}
   export interface RemoveLockUpType extends RemoveLockUpTypeParams {}
   export interface RemoveLockUpTypeMulti extends RemoveLockUpTypeMultiParams {}
+}
+
+interface AddLockUpToUserSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.AddLockUpToUser;
+  callback: EventCallback<LockUpTransferManagerAddLockUpToUserEventArgs_3_0_0>;
+}
+
+interface GetAddLockUpToUserLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.AddLockUpToUser;
+}
+
+interface RemoveLockUpFromUserSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.RemoveLockUpFromUser;
+  callback: EventCallback<LockUpTransferManagerRemoveLockUpFromUserEventArgs_3_0_0>;
+}
+
+interface GetRemoveLockUpFromUserLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.RemoveLockUpFromUser;
+}
+
+interface ModifyLockUpTypeSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.ModifyLockUpType;
+  callback: EventCallback<LockUpTransferManagerModifyLockUpTypeEventArgs_3_0_0>;
+}
+
+interface GetModifyLockUpTypeLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.ModifyLockUpType;
+}
+
+interface AddNewLockUpTypeSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.AddNewLockUpType;
+  callback: EventCallback<LockUpTransferManagerAddNewLockUpTypeEventArgs_3_0_0>;
+}
+
+interface GetAddNewLockUpTypeLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.AddNewLockUpType;
+}
+
+interface RemoveLockUpTypeSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.RemoveLockUpType;
+  callback: EventCallback<LockUpTransferManagerRemoveLockUpTypeEventArgs_3_0_0>;
+}
+
+interface GetRemoveLockUpTypeLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.RemoveLockUpType;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.Pause;
+  callback: EventCallback<LockUpTransferManagerPauseEventArgs_3_0_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.Unpause;
+  callback: EventCallback<LockUpTransferManagerUnpauseEventArgs_3_0_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: LockUpTransferManagerEvents_3_0_0.Unpause;
+}
+
+export interface LockUpTransferManagerSubscribeAsyncParams extends Subscribe {
+  (params: AddLockUpToUserSubscribeAsyncParams): Promise<string>;
+  (params: RemoveLockUpFromUserSubscribeAsyncParams): Promise<string>;
+  (params: ModifyLockUpTypeSubscribeAsyncParams): Promise<string>;
+  (params: AddNewLockUpTypeSubscribeAsyncParams): Promise<string>;
+  (params: RemoveLockUpTypeSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+}
+
+export interface GetLockUpTransferManagerLogsAsyncParams extends GetLogs {
+  (params: GetAddLockUpToUserLogsAsyncParams): Promise<
+    LogWithDecodedArgs<LockUpTransferManagerAddLockUpToUserEventArgs_3_0_0>[]
+  >;
+  (params: GetRemoveLockUpFromUserLogsAsyncParams): Promise<
+    LogWithDecodedArgs<LockUpTransferManagerRemoveLockUpFromUserEventArgs_3_0_0>[]
+  >;
+  (params: GetModifyLockUpTypeLogsAsyncParams): Promise<
+    LogWithDecodedArgs<LockUpTransferManagerModifyLockUpTypeEventArgs_3_0_0>[]
+  >;
+  (params: GetAddNewLockUpTypeLogsAsyncParams): Promise<
+    LogWithDecodedArgs<LockUpTransferManagerAddNewLockUpTypeEventArgs_3_0_0>[]
+  >;
+  (params: GetRemoveLockUpTypeLogsAsyncParams): Promise<
+    LogWithDecodedArgs<LockUpTransferManagerRemoveLockUpTypeEventArgs_3_0_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<LockUpTransferManagerPauseEventArgs_3_0_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<LockUpTransferManagerUnpauseEventArgs_3_0_0>[]>;
 }
 
 /**
@@ -791,4 +898,4 @@ export default abstract class LockUpTransferManagerCommon extends ModuleCommon {
 
 export function isLockUpTransferManager(wrapper: ContractWrapper): wrapper is LockUpTransferManagerCommon {
   return wrapper instanceof LockUpTransferManagerCommon;
-};
+}
