+++ src/contract_wrappers/modules/transfer_manager/manual_approval_transfer_manager_wrapper/3.0.0.ts
@@ -1,81 +1,23 @@
-import { ManualApprovalTransferManagerContract_3_0_0, ManualApprovalTransferManagerEvents_3_0_0, ManualApprovalTransferManagerAddManualApprovalEventArgs_3_0_0, ManualApprovalTransferManagerModifyManualApprovalEventArgs_3_0_0, ManualApprovalTransferManagerRevokeManualApprovalEventArgs_3_0_0, ManualApprovalTransferManagerPauseEventArgs_3_0_0, ManualApprovalTransferManagerUnpauseEventArgs_3_0_0, LogWithDecodedArgs, ManualApprovalTransferManagerEventArgs_3_0_0, Web3Wrapper } from '@polymathnetwork/abi-wrappers';
+import {
+  ManualApprovalTransferManagerContract_3_0_0,
+  ManualApprovalTransferManagerEvents_3_0_0,
+  LogWithDecodedArgs,
+  ManualApprovalTransferManagerEventArgs_3_0_0,
+  Web3Wrapper,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import ManualApprovalTransferManagerCommon from './common';
-import { ContractVersion, Constructor, SubscribeAsyncParams, GetLogsAsyncParams, Subscribe, GetLogs, EventCallback } from '../../../../types';
+import ManualApprovalTransferManagerCommon, {
+  ManualApprovalTransferManagerSubscribeAsyncParams,
+  GetManualApprovalTransferManagerLogsAsyncParams,
+} from './common';
+import { ContractVersion, Constructor, SubscribeAsyncParams, GetLogsAsyncParams } from '../../../../types';
 import { WithModule_3_0_0 } from '../../module_wrapper';
 import assert from '../../../../utils/assert';
 import ContractFactory from '../../../../factories/contractFactory';
 
-interface AddManualApprovalSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.AddManualApproval;
-  callback: EventCallback<ManualApprovalTransferManagerAddManualApprovalEventArgs_3_0_0>;
-}
-
-interface GetAddManualApprovalLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.AddManualApproval;
-}
-
-interface ModifyManualApprovalSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.ModifyManualApproval;
-  callback: EventCallback<ManualApprovalTransferManagerModifyManualApprovalEventArgs_3_0_0>;
-}
-
-interface GetModifyManualApprovalLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.ModifyManualApproval;
-}
-
-interface RevokeManualApprovalSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.RevokeManualApproval;
-  callback: EventCallback<ManualApprovalTransferManagerRevokeManualApprovalEventArgs_3_0_0>;
-}
-
-interface GetRevokeManualApprovalLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.RevokeManualApproval;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.Pause;
-  callback: EventCallback<ManualApprovalTransferManagerPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.Unpause;
-  callback: EventCallback<ManualApprovalTransferManagerUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: ManualApprovalTransferManagerEvents_3_0_0.Unpause;
-}
-
-interface ManualApprovalTransferManagerSubscribeAsyncParams extends Subscribe {
-  (params: AddManualApprovalSubscribeAsyncParams): Promise<string>;
-  (params: ModifyManualApprovalSubscribeAsyncParams): Promise<string>;
-  (params: RevokeManualApprovalSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetManualApprovalTransferManagerLogsAsyncParams extends GetLogs {
-  (params: GetAddManualApprovalLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ManualApprovalTransferManagerAddManualApprovalEventArgs_3_0_0>[]
-  >;
-  (params: GetModifyManualApprovalLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ManualApprovalTransferManagerModifyManualApprovalEventArgs_3_0_0>[]
-  >;
-  (params: GetRevokeManualApprovalLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ManualApprovalTransferManagerRevokeManualApprovalEventArgs_3_0_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<ManualApprovalTransferManagerPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<
-    LogWithDecodedArgs<ManualApprovalTransferManagerUnpauseEventArgs_3_0_0>[]
-  >;
-}
-
-const ManualApprovalTransferManagerBase_3_0_0 = WithModule_3_0_0(ManualApprovalTransferManagerCommon as unknown as Constructor<ManualApprovalTransferManagerCommon>);
+const ManualApprovalTransferManagerBase_3_0_0 = WithModule_3_0_0(
+  (ManualApprovalTransferManagerCommon as unknown) as Constructor<ManualApprovalTransferManagerCommon>,
+);
 
 export class ManualApprovalTransferManager_3_0_0 extends ManualApprovalTransferManagerBase_3_0_0 {
   public contract: Promise<ManualApprovalTransferManagerContract_3_0_0>;
