+++ src/contract_wrappers/modules/transfer_manager/restricted_partial_sale_transfer_manager_wrapper/common.ts
@@ -2,17 +2,23 @@ import {
   RestrictedPartialSaleTMContract_3_1_0,
   Web3Wrapper,
   BigNumber,
+  RestrictedPartialSaleTMEvents_3_1_0,
+  RestrictedPartialSaleTMChangedExemptWalletListEventArgs_3_1_0,
+  RestrictedPartialSaleTMPauseEventArgs_3_1_0,
+  RestrictedPartialSaleTMUnpauseEventArgs_3_1_0,
+  LogWithDecodedArgs,
 } from '@polymathnetwork/abi-wrappers';
-import {
-  parsePermBytes32Value,
-  parseTransferResult,
-  valueToWei,
-} from '../../../../utils/convert';
+import { parsePermBytes32Value, parseTransferResult, valueToWei } from '../../../../utils/convert';
 import ContractFactory from '../../../../factories/contractFactory';
 import {
   ErrorCode,
   Perm,
   TxParams,
+  EventCallback,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  Subscribe,
+  GetLogs,
 } from '../../../../types';
 import { ModuleCommon } from '../../module_wrapper';
 import assert from '../../../../utils/assert';
@@ -27,6 +33,47 @@ export namespace RestrictedPartialSaleTransferManagerTransactionParams {
   export interface ChangeExemptWalletList extends ChangeExemptWalletListParams {}
 }
 
+interface ChangedExemptWalletListSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.ChangedExemptWalletList;
+  callback: EventCallback<RestrictedPartialSaleTMChangedExemptWalletListEventArgs_3_1_0>;
+}
+
+interface GetChangedExemptWalletListLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.ChangedExemptWalletList;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.Pause;
+  callback: EventCallback<RestrictedPartialSaleTMPauseEventArgs_3_1_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.Unpause;
+  callback: EventCallback<RestrictedPartialSaleTMUnpauseEventArgs_3_1_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: RestrictedPartialSaleTMEvents_3_1_0.Unpause;
+}
+
+export interface RestrictedPartialSaleTransferManagerSubscribeAsyncParams extends Subscribe {
+  (params: ChangedExemptWalletListSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+}
+
+export interface GetRestrictedPartialSaleTransferManagerLogsAsyncParams extends GetLogs {
+  (params: GetChangedExemptWalletListLogsAsyncParams): Promise<
+    LogWithDecodedArgs<RestrictedPartialSaleTMChangedExemptWalletListEventArgs_3_1_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<RestrictedPartialSaleTMPauseEventArgs_3_1_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<RestrictedPartialSaleTMUnpauseEventArgs_3_1_0>[]>;
+}
+
 /**
  * @param wallet Ethereum wallet/contract address that need to be exempted
  * @param exempted Boolean value used to add (i.e true) or remove (i.e false) from the list
@@ -35,7 +82,6 @@ interface ChangeExemptWalletListParams extends TxParams {
   wallet: string;
   exempted: boolean;
 }
-
 
 /**
  * @param wallets Ethereum wallet/contract addresses that need to be exempted
@@ -167,6 +213,8 @@ export default abstract class RestrictedPartialSaleTransferManagerCommon extends
   };
 }
 
-export function isRestrictedPartialSaleTransferManager(wrapper: ContractWrapper): wrapper is RestrictedPartialSaleTransferManagerCommon {
+export function isRestrictedPartialSaleTransferManager(
+  wrapper: ContractWrapper,
+): wrapper is RestrictedPartialSaleTransferManagerCommon {
   return wrapper instanceof RestrictedPartialSaleTransferManagerCommon;
-};
+}
