+++ src/contract_wrappers/modules/transfer_manager/restricted_partial_sale_transfer_manager_wrapper/index.ts
@@ -1,8 +1,18 @@
 /* istanbul ignore file */
+import { RestrictedPartialSaleTMEventArgs_3_1_0 } from '@polymathnetwork/abi-wrappers';
 import { RestrictedPartialSaleTransferManager_3_1_0, isRestrictedPartialSaleTransferManager_3_1_0 } from './3.1.0';
 
 import Common, { isRestrictedPartialSaleTransferManager } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../../types';
+
+export type RestrictedPartialSaleTransferManagerEventArgs = RestrictedPartialSaleTMEventArgs_3_1_0;
+
+export {
+  RestrictedPartialSaleTMEvents_3_1_0 as RestrictedPartialSaleTransferManagerEvents,
+  RestrictedPartialSaleTMChangedExemptWalletListEventArgs_3_1_0 as RestrictedPartialSaleTMChangedExemptWalletListEventArgs,
+  RestrictedPartialSaleTMPauseEventArgs_3_1_0 as RestrictedPartialSaleTMPauseEventArgs,
+  RestrictedPartialSaleTMUnpauseEventArgs_3_1_0 as RestrictedPartialSaleTMUnpauseEventArgs,
+} from '@polymathnetwork/abi-wrappers';
 
 export type RestrictedPartialSaleTransferManager = RestrictedPartialSaleTransferManager_3_1_0;
