+++ src/contract_wrappers/modules/transfer_manager/volume_restriction_transfer_manager_wrapper/3.0.0.ts
@@ -1,271 +1,85 @@
-import { VolumeRestrictionTMEvents_3_0_0, VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0, VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0, VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0, VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0, VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0, VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0, VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0, VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0, VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0, VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0, VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0, VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0, VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0, VolumeRestrictionTMPauseEventArgs_3_0_0, VolumeRestrictionTMUnpauseEventArgs_3_0_0, LogWithDecodedArgs, VolumeRestrictionTMContract_3_0_0, Web3Wrapper, VolumeRestrictionTMEventArgs_3_0_0 } from '@polymathnetwork/abi-wrappers';
+import {
+  VolumeRestrictionTMEvents_3_0_0,
+  LogWithDecodedArgs,
+  VolumeRestrictionTMContract_3_0_0,
+  Web3Wrapper,
+  VolumeRestrictionTMEventArgs_3_0_0,
+} from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import VolumeRestrictionTransferManagerCommon from './common';
-import { ContractVersion, Constructor, SubscribeAsyncParams, EventCallback, GetLogsAsyncParams, Subscribe, GetLogs } from '../../../../types';
+import VolumeRestrictionTransferManagerCommon, {
+  VolumeRestrictionTransferManagerSubscribeAsyncParams,
+  GetVolumeRestrictionTransferManagerLogsAsyncParams,
+} from './common';
+import { ContractVersion, Constructor, SubscribeAsyncParams, GetLogsAsyncParams } from '../../../../types';
 import { WithModule_3_0_0 } from '../../module_wrapper';
 import ContractFactory from '../../../../factories/contractFactory';
 import assert from '../../../../utils/assert';
 
-interface ChangedExemptWalletListSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ChangedExemptWalletList;
-  callback: EventCallback<VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0>;
-}
-
-interface GetChangedExemptWalletListLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ChangedExemptWalletList;
-}
-
-interface AddIndividualRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualRestriction;
-  callback: EventCallback<VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0>;
-}
-
-interface GetAddIndividualRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualRestriction;
-}
-
-interface AddIndividualDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualDailyRestriction;
-  callback: EventCallback<VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0>;
-}
-
-interface GetAddIndividualDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualDailyRestriction;
-}
-
-interface ModifyIndividualRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualRestriction;
-  callback: EventCallback<VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0>;
-}
-
-interface GetModifyIndividualRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualRestriction;
-}
-
-interface ModifyIndividualDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualDailyRestriction;
-  callback: EventCallback<VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0>;
-}
-
-interface GetModifyIndividualDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualDailyRestriction;
-}
-
-interface AddDefaultRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultRestriction;
-  callback: EventCallback<VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0>;
-}
-
-interface GetAddDefaultRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultRestriction;
-}
-
-interface AddDefaultDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultDailyRestriction;
-  callback: EventCallback<VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0>;
-}
-
-interface GetAddDefaultDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultDailyRestriction;
-}
-
-interface ModifyDefaultRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultRestriction;
-  callback: EventCallback<VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0>;
-}
-
-interface GetModifyDefaultRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultRestriction;
-}
-
-interface ModifyDefaultDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultDailyRestriction;
-  callback: EventCallback<VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0>;
-}
-
-interface GetModifyDefaultDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultDailyRestriction;
-}
-
-interface IndividualRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualRestrictionRemoved;
-  callback: EventCallback<VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0>;
-}
-
-interface GetIndividualRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualRestrictionRemoved;
-}
-
-interface IndividualDailyRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualDailyRestrictionRemoved;
-  callback: EventCallback<VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0>;
-}
-
-interface GetIndividualDailyRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualDailyRestrictionRemoved;
-}
-
-interface DefaultRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultRestrictionRemoved;
-  callback: EventCallback<VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0>;
-}
-
-interface GetDefaultRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultRestrictionRemoved;
-}
-
-interface DefaultDailyRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultDailyRestrictionRemoved;
-  callback: EventCallback<VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0>;
-}
-
-interface GetDefaultDailyRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultDailyRestrictionRemoved;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.Pause;
-  callback: EventCallback<VolumeRestrictionTMPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.Unpause;
-  callback: EventCallback<VolumeRestrictionTMUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VolumeRestrictionTMEvents_3_0_0.Unpause;
-}
-
-interface VolumeRestrictionTransferManagerSubscribeAsyncParams extends Subscribe {
-  (params: ChangedExemptWalletListSubscribeAsyncParams): Promise<string>;
-  (params: AddIndividualRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: AddIndividualDailyRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: ModifyIndividualRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: ModifyIndividualDailyRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: AddDefaultRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: AddDefaultDailyRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: ModifyDefaultRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: ModifyDefaultDailyRestrictionSubscribeAsyncParams): Promise<string>;
-  (params: IndividualRestrictionRemovedSubscribeAsyncParams): Promise<string>;
-  (params: IndividualDailyRestrictionRemovedSubscribeAsyncParams): Promise<string>;
-  (params: DefaultRestrictionRemovedSubscribeAsyncParams): Promise<string>;
-  (params: DefaultDailyRestrictionRemovedSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetVolumeRestrictionTransferManagerLogsAsyncParams extends GetLogs {
-  (params: GetChangedExemptWalletListLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0>[]
-  >;
-  (params: GetAddIndividualRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetAddIndividualDailyRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetModifyIndividualRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetModifyIndividualDailyRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetAddDefaultRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetAddDefaultDailyRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetModifyDefaultRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetModifyDefaultDailyRestrictionLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0>[]
-  >;
-  (params: GetIndividualRestrictionRemovedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0>[]
-  >;
-  (params: GetIndividualDailyRestrictionRemovedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0>[]
-  >;
-  (params: GetDefaultRestrictionRemovedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0>[]
-  >;
-  (params: GetDefaultDailyRestrictionRemovedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<VolumeRestrictionTMPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<VolumeRestrictionTMUnpauseEventArgs_3_0_0>[]>;
-}
-
-const VolumeRestrictionTransferManagerBase_3_0_0 = WithModule_3_0_0(VolumeRestrictionTransferManagerCommon as unknown as Constructor<VolumeRestrictionTransferManagerCommon>);
+const VolumeRestrictionTransferManagerBase_3_0_0 = WithModule_3_0_0(
+  (VolumeRestrictionTransferManagerCommon as unknown) as Constructor<VolumeRestrictionTransferManagerCommon>,
+);
 
 export class VolumeRestrictionTransferManager_3_0_0 extends VolumeRestrictionTransferManagerBase_3_0_0 {
   public contract: Promise<VolumeRestrictionTMContract_3_0_0>;
 
   public contractVersion = ContractVersion.V3_0_0;
-  
-    /**
-     * Instantiate VolumeRestrictionManagerWrapper
-     * @param web3Wrapper Web3Wrapper instance to use
-     * @param contract
-     */
-    public constructor(
-      web3Wrapper: Web3Wrapper,
-      contract: Promise<VolumeRestrictionTMContract_3_0_0>,
-      contractFactory: ContractFactory,
-    ) {
-      super(web3Wrapper, contract, contractFactory);
-      this.contract = contract;
-    }
 
-    /**
-     * Subscribe to an event type emitted by the contract.
-     * @return Subscription token used later to unsubscribe
-     */
-    public subscribeAsync: VolumeRestrictionTransferManagerSubscribeAsyncParams = async <
-      ArgsType extends VolumeRestrictionTMEventArgs_3_0_0
-    >(
-      params: SubscribeAsyncParams,
-    ): Promise<string> => {
-      assert.doesBelongToStringEnum('eventName', params.eventName, VolumeRestrictionTMEvents_3_0_0);
-      assert.doesConformToSchema('indexFilterValues', params.indexFilterValues, schemas.indexFilterValuesSchema);
-      assert.isFunction('callback', params.callback);
-      const normalizedContractAddress = (await this.contract).address.toLowerCase();
-      const subscriptionToken = await this.subscribeInternal<ArgsType>(
-        normalizedContractAddress,
-        params.eventName,
-        params.indexFilterValues,
-        params.callback,
-        params.isVerbose,
-      );
-      return subscriptionToken;
-    };
-  
-    /**
-     * Gets historical logs without creating a subscription
-     * @return Array of logs that match the parameters
-     */
-    public getLogsAsync: GetVolumeRestrictionTransferManagerLogsAsyncParams = async <
-      ArgsType extends VolumeRestrictionTMEventArgs_3_0_0
-    >(
-      params: GetLogsAsyncParams,
-    ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
-      assert.doesBelongToStringEnum('eventName', params.eventName, VolumeRestrictionTMEvents_3_0_0);
-      const normalizedContractAddress = (await this.contract).address.toLowerCase();
-      const logs = await this.getLogsAsyncInternal<ArgsType>(
-        normalizedContractAddress,
-        params.eventName,
-        params.blockRange,
-        params.indexFilterValues,
-      );
-      return logs;
-    };
+  /**
+   * Instantiate VolumeRestrictionManagerWrapper
+   * @param web3Wrapper Web3Wrapper instance to use
+   * @param contract
+   */
+  public constructor(
+    web3Wrapper: Web3Wrapper,
+    contract: Promise<VolumeRestrictionTMContract_3_0_0>,
+    contractFactory: ContractFactory,
+  ) {
+    super(web3Wrapper, contract, contractFactory);
+    this.contract = contract;
+  }
+
+  /**
+   * Subscribe to an event type emitted by the contract.
+   * @return Subscription token used later to unsubscribe
+   */
+  public subscribeAsync: VolumeRestrictionTransferManagerSubscribeAsyncParams = async <
+    ArgsType extends VolumeRestrictionTMEventArgs_3_0_0
+  >(
+    params: SubscribeAsyncParams,
+  ): Promise<string> => {
+    assert.doesBelongToStringEnum('eventName', params.eventName, VolumeRestrictionTMEvents_3_0_0);
+    assert.doesConformToSchema('indexFilterValues', params.indexFilterValues, schemas.indexFilterValuesSchema);
+    assert.isFunction('callback', params.callback);
+    const normalizedContractAddress = (await this.contract).address.toLowerCase();
+    const subscriptionToken = await this.subscribeInternal<ArgsType>(
+      normalizedContractAddress,
+      params.eventName,
+      params.indexFilterValues,
+      params.callback,
+      params.isVerbose,
+    );
+    return subscriptionToken;
+  };
+
+  /**
+   * Gets historical logs without creating a subscription
+   * @return Array of logs that match the parameters
+   */
+  public getLogsAsync: GetVolumeRestrictionTransferManagerLogsAsyncParams = async <
+    ArgsType extends VolumeRestrictionTMEventArgs_3_0_0
+  >(
+    params: GetLogsAsyncParams,
+  ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
+    assert.doesBelongToStringEnum('eventName', params.eventName, VolumeRestrictionTMEvents_3_0_0);
+    const normalizedContractAddress = (await this.contract).address.toLowerCase();
+    const logs = await this.getLogsAsyncInternal<ArgsType>(
+      normalizedContractAddress,
+      params.eventName,
+      params.blockRange,
+      params.indexFilterValues,
+    );
+    return logs;
+  };
 }
 
 export function isVolumeRestrictionTransferManager_3_0_0(
