+++ src/contract_wrappers/modules/transfer_manager/volume_restriction_transfer_manager_wrapper/common.ts
@@ -2,7 +2,24 @@ import {
   VolumeRestrictionTMContract_3_0_0,
   Web3Wrapper,
   BigNumber,
-  PolyResponse
+  PolyResponse,
+  VolumeRestrictionTMEvents_3_0_0,
+  VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0,
+  VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0,
+  VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0,
+  VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0,
+  VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0,
+  VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0,
+  VolumeRestrictionTMPauseEventArgs_3_0_0,
+  VolumeRestrictionTMUnpauseEventArgs_3_0_0,
+  LogWithDecodedArgs,
 } from '@polymathnetwork/abi-wrappers';
 import assert from '../../../../utils/assert';
 import { ModuleCommon } from '../../module_wrapper';
@@ -23,7 +40,12 @@ import {
   RestrictionType,
   PERCENTAGE_DECIMALS,
   ErrorCode,
-  TransferResult
+  TransferResult,
+  EventCallback,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  Subscribe,
+  GetLogs,
 } from '../../../../types';
 import ContractWrapper from '../../../contract_wrapper';
 
@@ -45,6 +67,203 @@ export namespace VolumeRestrictionTransferManagerTransactionParams {
   export interface ModifyIndividualDailyRestrictionMulti extends IndividualDailyRestrictionMultiParams {}
   export interface AddIndividualRestrictionMulti extends IndividualRestrictionMultiParams {}
   export interface ModifyIndividualRestrictionMulti extends IndividualRestrictionMultiParams {}
+}
+
+interface ChangedExemptWalletListSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ChangedExemptWalletList;
+  callback: EventCallback<VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0>;
+}
+
+interface GetChangedExemptWalletListLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ChangedExemptWalletList;
+}
+
+interface AddIndividualRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualRestriction;
+  callback: EventCallback<VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0>;
+}
+
+interface GetAddIndividualRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualRestriction;
+}
+
+interface AddIndividualDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualDailyRestriction;
+  callback: EventCallback<VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0>;
+}
+
+interface GetAddIndividualDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddIndividualDailyRestriction;
+}
+
+interface ModifyIndividualRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualRestriction;
+  callback: EventCallback<VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0>;
+}
+
+interface GetModifyIndividualRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualRestriction;
+}
+
+interface ModifyIndividualDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualDailyRestriction;
+  callback: EventCallback<VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0>;
+}
+
+interface GetModifyIndividualDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyIndividualDailyRestriction;
+}
+
+interface AddDefaultRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultRestriction;
+  callback: EventCallback<VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0>;
+}
+
+interface GetAddDefaultRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultRestriction;
+}
+
+interface AddDefaultDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultDailyRestriction;
+  callback: EventCallback<VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0>;
+}
+
+interface GetAddDefaultDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.AddDefaultDailyRestriction;
+}
+
+interface ModifyDefaultRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultRestriction;
+  callback: EventCallback<VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0>;
+}
+
+interface GetModifyDefaultRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultRestriction;
+}
+
+interface ModifyDefaultDailyRestrictionSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultDailyRestriction;
+  callback: EventCallback<VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0>;
+}
+
+interface GetModifyDefaultDailyRestrictionLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.ModifyDefaultDailyRestriction;
+}
+
+interface IndividualRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualRestrictionRemoved;
+  callback: EventCallback<VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0>;
+}
+
+interface GetIndividualRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualRestrictionRemoved;
+}
+
+interface IndividualDailyRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualDailyRestrictionRemoved;
+  callback: EventCallback<VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0>;
+}
+
+interface GetIndividualDailyRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.IndividualDailyRestrictionRemoved;
+}
+
+interface DefaultRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultRestrictionRemoved;
+  callback: EventCallback<VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0>;
+}
+
+interface GetDefaultRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultRestrictionRemoved;
+}
+
+interface DefaultDailyRestrictionRemovedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultDailyRestrictionRemoved;
+  callback: EventCallback<VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0>;
+}
+
+interface GetDefaultDailyRestrictionRemovedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.DefaultDailyRestrictionRemoved;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.Pause;
+  callback: EventCallback<VolumeRestrictionTMPauseEventArgs_3_0_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.Unpause;
+  callback: EventCallback<VolumeRestrictionTMUnpauseEventArgs_3_0_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VolumeRestrictionTMEvents_3_0_0.Unpause;
+}
+
+export interface VolumeRestrictionTransferManagerSubscribeAsyncParams extends Subscribe {
+  (params: ChangedExemptWalletListSubscribeAsyncParams): Promise<string>;
+  (params: AddIndividualRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: AddIndividualDailyRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: ModifyIndividualRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: ModifyIndividualDailyRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: AddDefaultRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: AddDefaultDailyRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: ModifyDefaultRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: ModifyDefaultDailyRestrictionSubscribeAsyncParams): Promise<string>;
+  (params: IndividualRestrictionRemovedSubscribeAsyncParams): Promise<string>;
+  (params: IndividualDailyRestrictionRemovedSubscribeAsyncParams): Promise<string>;
+  (params: DefaultRestrictionRemovedSubscribeAsyncParams): Promise<string>;
+  (params: DefaultDailyRestrictionRemovedSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+}
+
+export interface GetVolumeRestrictionTransferManagerLogsAsyncParams extends GetLogs {
+  (params: GetChangedExemptWalletListLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0>[]
+  >;
+  (params: GetAddIndividualRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetAddIndividualDailyRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetModifyIndividualRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetModifyIndividualDailyRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetAddDefaultRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetAddDefaultDailyRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetModifyDefaultRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetModifyDefaultDailyRestrictionLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0>[]
+  >;
+  (params: GetIndividualRestrictionRemovedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0>[]
+  >;
+  (params: GetIndividualDailyRestrictionRemovedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0>[]
+  >;
+  (params: GetDefaultRestrictionRemovedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0>[]
+  >;
+  (params: GetDefaultDailyRestrictionRemovedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<VolumeRestrictionTMPauseEventArgs_3_0_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<VolumeRestrictionTMUnpauseEventArgs_3_0_0>[]>;
 }
 
 /**
@@ -408,7 +627,9 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   /**
    * Use to add the new individual daily restriction for multiple token holders
    */
-  public addIndividualDailyRestrictionMulti = async (params: IndividualDailyRestrictionMultiParams): Promise<PolyResponse> => {
+  public addIndividualDailyRestrictionMulti = async (
+    params: IndividualDailyRestrictionMultiParams,
+  ): Promise<PolyResponse> => {
     assert.assert(
       await this.isCallerAllowed(params.txData, Perm.Admin),
       ErrorCode.Unauthorized,
@@ -576,7 +797,9 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   /**
    * Use to remove the individual restriction for a given address
    */
-  public removeIndividualRestrictionMulti = async (params: RemoveIndividualRestrictionMultiParams): Promise<PolyResponse> => {
+  public removeIndividualRestrictionMulti = async (
+    params: RemoveIndividualRestrictionMultiParams,
+  ): Promise<PolyResponse> => {
     assert.assert(
       await this.isCallerAllowed(params.txData, Perm.Admin),
       ErrorCode.Unauthorized,
@@ -601,7 +824,9 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   /**
    * Use to remove the individual daily restriction for a given address
    */
-  public removeIndividualDailyRestriction = async (params: HolderIndividualRestrictionParams): Promise<PolyResponse> => {
+  public removeIndividualDailyRestriction = async (
+    params: HolderIndividualRestrictionParams,
+  ): Promise<PolyResponse> => {
     assert.assert(
       await this.isCallerAllowed(params.txData, Perm.Admin),
       ErrorCode.Unauthorized,
@@ -622,7 +847,9 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   /**
    * Use to remove the individual daily restriction for a given address
    */
-  public removeIndividualDailyRestrictionMulti = async (params: RemoveIndividualRestrictionMultiParams): Promise<PolyResponse> => {
+  public removeIndividualDailyRestrictionMulti = async (
+    params: RemoveIndividualRestrictionMultiParams,
+  ): Promise<PolyResponse> => {
     assert.assert(
       await this.isCallerAllowed(params.txData, Perm.Admin),
       ErrorCode.Unauthorized,
@@ -729,7 +956,9 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   /**
    * Use to modify the existing individual daily restriction for multiple token holders
    */
-  public modifyIndividualDailyRestrictionMulti = async (params: IndividualDailyRestrictionMultiParams): Promise<PolyResponse> => {
+  public modifyIndividualDailyRestrictionMulti = async (
+    params: IndividualDailyRestrictionMultiParams,
+  ): Promise<PolyResponse> => {
     assert.assert(
       await this.isCallerAllowed(params.txData, Perm.Admin),
       ErrorCode.Unauthorized,
@@ -985,7 +1214,8 @@ export default abstract class VolumeRestrictionTransferManagerCommon extends Mod
   };
 }
 
-export function isVolumeRestrictionTransferManager(wrapper: ContractWrapper): wrapper is VolumeRestrictionTransferManagerCommon {
+export function isVolumeRestrictionTransferManager(
+  wrapper: ContractWrapper,
+): wrapper is VolumeRestrictionTransferManagerCommon {
   return wrapper instanceof VolumeRestrictionTransferManagerCommon;
-};
-  
+}
