+++ src/contract_wrappers/modules/wallet/vesting_escrow_wallet_wrapper/3.1.0.ts
@@ -1,23 +1,11 @@
 import {
   VestingEscrowWalletContract_3_1_0,
+  VestingEscrowWalletEventArgs_3_1_0,
+  VestingEscrowWalletEvents_3_1_0,
   Web3Wrapper,
   BigNumber,
   PolyResponse,
   LogWithDecodedArgs,
-  VestingEscrowWalletEventArgs_3_1_0,
-  VestingEscrowWalletEvents_3_1_0,
-  VestingEscrowWalletAddScheduleEventArgs_3_1_0,
-  VestingEscrowWalletModifyScheduleEventArgs_3_1_0,
-  VestingEscrowWalletRevokeAllSchedulesEventArgs_3_1_0,
-  VestingEscrowWalletRevokeScheduleEventArgs_3_1_0,
-  VestingEscrowWalletDepositTokensEventArgs_3_1_0,
-  VestingEscrowWalletSendToTreasuryEventArgs_3_1_0,
-  VestingEscrowWalletSendTokensEventArgs_3_1_0,
-  VestingEscrowWalletAddTemplateEventArgs_3_1_0,
-  VestingEscrowWalletRemoveTemplateEventArgs_3_1_0,
-  VestingEscrowWalletTreasuryWalletChangedEventArgs_3_1_0,
-  VestingEscrowWalletPauseEventArgs_3_1_0,
-  VestingEscrowWalletUnpauseEventArgs_3_1_0,
 } from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
 import assert from '../../../../utils/assert';
@@ -29,166 +17,14 @@ import {
   Perm,
   GetLogsAsyncParams,
   SubscribeAsyncParams,
-  EventCallback,
-  Subscribe,
-  GetLogs,
   Constructor,
 } from '../../../../types';
 import { weiToValue, stringToBytes32, dateToBigNumber } from '../../../../utils/convert';
-import VestingEscrowWalletCommon from './common';
+import VestingEscrowWalletCommon, {
+  VestingEscrowWalletSubscribeAsyncParams,
+  GetVestingEscrowWalletLogsAsyncParams,
+} from './common';
 import { WithModule_3_0_0 } from '../../module_wrapper';
-
-interface AddScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.AddSchedule;
-  callback: EventCallback<VestingEscrowWalletAddScheduleEventArgs_3_1_0>;
-}
-
-interface GetAddScheduleLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.AddSchedule;
-}
-
-interface ModifyScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.ModifySchedule;
-  callback: EventCallback<VestingEscrowWalletModifyScheduleEventArgs_3_1_0>;
-}
-
-interface GetModifyScheduleLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.ModifySchedule;
-}
-
-interface RevokeAllSchedulesSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RevokeAllSchedules;
-  callback: EventCallback<VestingEscrowWalletRevokeAllSchedulesEventArgs_3_1_0>;
-}
-
-interface GetRevokeAllSchedulesLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RevokeAllSchedules;
-}
-
-interface RevokeScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RevokeSchedule;
-  callback: EventCallback<VestingEscrowWalletRevokeScheduleEventArgs_3_1_0>;
-}
-
-interface GetRevokeScheduleLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RevokeSchedule;
-}
-
-interface DepositTokensSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.DepositTokens;
-  callback: EventCallback<VestingEscrowWalletDepositTokensEventArgs_3_1_0>;
-}
-
-interface GetDepositTokensLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.DepositTokens;
-}
-
-interface SendToTreasurySubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.SendToTreasury;
-  callback: EventCallback<VestingEscrowWalletSendToTreasuryEventArgs_3_1_0>;
-}
-
-interface GetSendToTreasuryLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.SendToTreasury;
-}
-
-interface SendTokensSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.SendTokens;
-  callback: EventCallback<VestingEscrowWalletSendTokensEventArgs_3_1_0>;
-}
-
-interface GetSendTokensLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.SendTokens;
-}
-
-interface AddTemplateSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.AddTemplate;
-  callback: EventCallback<VestingEscrowWalletAddTemplateEventArgs_3_1_0>;
-}
-
-interface GetAddTemplateLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.AddTemplate;
-}
-
-interface RemoveTemplateSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RemoveTemplate;
-  callback: EventCallback<VestingEscrowWalletRemoveTemplateEventArgs_3_1_0>;
-}
-
-interface GetRemoveTemplateLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.RemoveTemplate;
-}
-
-interface TreasuryWalletChangedSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.TreasuryWalletChanged;
-  callback: EventCallback<VestingEscrowWalletTreasuryWalletChangedEventArgs_3_1_0>;
-}
-
-interface GetTreasuryWalletChangedLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.TreasuryWalletChanged;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.Pause;
-  callback: EventCallback<VestingEscrowWalletPauseEventArgs_3_1_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.Unpause;
-  callback: EventCallback<VestingEscrowWalletUnpauseEventArgs_3_1_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: VestingEscrowWalletEvents_3_1_0.Unpause;
-}
-
-interface VestingEscrowWalletSubscribeAsyncParams extends Subscribe {
-  (params: AddScheduleSubscribeAsyncParams): Promise<string>;
-  (params: ModifyScheduleSubscribeAsyncParams): Promise<string>;
-  (params: RevokeAllSchedulesSubscribeAsyncParams): Promise<string>;
-  (params: RevokeScheduleSubscribeAsyncParams): Promise<string>;
-  (params: DepositTokensSubscribeAsyncParams): Promise<string>;
-  (params: SendToTreasurySubscribeAsyncParams): Promise<string>;
-  (params: SendTokensSubscribeAsyncParams): Promise<string>;
-  (params: AddTemplateSubscribeAsyncParams): Promise<string>;
-  (params: RemoveTemplateSubscribeAsyncParams): Promise<string>;
-  (params: TreasuryWalletChangedSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetVestingEscrowWalletLogsAsyncParams extends GetLogs {
-  (params: GetAddScheduleLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletAddScheduleEventArgs_3_1_0>[]>;
-  (params: GetModifyScheduleLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletModifyScheduleEventArgs_3_1_0>[]
-  >;
-  (params: GetRevokeAllSchedulesLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletRevokeAllSchedulesEventArgs_3_1_0>[]
-  >;
-  (params: GetRevokeScheduleLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletRevokeScheduleEventArgs_3_1_0>[]
-  >;
-  (params: GetDepositTokensLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletDepositTokensEventArgs_3_1_0>[]
-  >;
-  (params: GetSendToTreasuryLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletSendToTreasuryEventArgs_3_1_0>[]
-  >;
-  (params: GetSendTokensLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletSendTokensEventArgs_3_1_0>[]>;
-  (params: GetAddTemplateLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletAddTemplateEventArgs_3_1_0>[]>;
-  (params: GetRemoveTemplateLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletRemoveTemplateEventArgs_3_1_0>[]
-  >;
-  (params: GetTreasuryWalletChangedLogsAsyncParams): Promise<
-    LogWithDecodedArgs<VestingEscrowWalletTreasuryWalletChangedEventArgs_3_1_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletPauseEventArgs_3_1_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletUnpauseEventArgs_3_1_0>[]>;
-}
 
 /**
  * @param beneficiary Address of the beneficiary for whom it is modified
@@ -211,7 +47,9 @@ interface Template {
 }
 
 // uses 3.0.0 Module contract
-const VestingEscrowWalletBase_3_1_0 = WithModule_3_0_0(VestingEscrowWalletCommon as unknown as Constructor<VestingEscrowWalletCommon>);
+const VestingEscrowWalletBase_3_1_0 = WithModule_3_0_0((VestingEscrowWalletCommon as unknown) as Constructor<
+  VestingEscrowWalletCommon
+>);
 
 /**
  * This class includes the functionality related to interacting with the Vesting Escrow Wallet contract.
