+++ src/contract_wrappers/modules/wallet/vesting_escrow_wallet_wrapper/common.ts
@@ -1,14 +1,39 @@
 import {
   VestingEscrowWalletContract_3_0_0,
+  VestingEscrowWalletContract_3_1_0,
+  VestingEscrowWalletAddScheduleEventArgs_3_0_0,
+  VestingEscrowWalletAddTemplateEventArgs_3_0_0,
+  VestingEscrowWalletDepositTokensEventArgs_3_0_0,
+  VestingEscrowWalletModifyScheduleEventArgs_3_0_0,
+  VestingEscrowWalletPauseEventArgs_3_0_0,
+  VestingEscrowWalletRemoveTemplateEventArgs_3_0_0,
+  VestingEscrowWalletRevokeAllSchedulesEventArgs_3_0_0,
+  VestingEscrowWalletRevokeScheduleEventArgs_3_0_0,
+  VestingEscrowWalletSendToTreasuryEventArgs_3_0_0,
+  VestingEscrowWalletSendTokensEventArgs_3_0_0,
+  VestingEscrowWalletTreasuryWalletChangedEventArgs_3_0_0,
+  VestingEscrowWalletUnpauseEventArgs_3_0_0,
   Web3Wrapper,
   BigNumber,
-  VestingEscrowWalletContract_3_1_0,
   PolyResponse,
+  LogWithDecodedArgs,
+  VestingEscrowWalletEvents_3_0_0,
 } from '@polymathnetwork/abi-wrappers';
 import assert from '../../../../utils/assert';
 import { ModuleCommon } from '../../module_wrapper';
 import ContractFactory from '../../../../factories/contractFactory';
-import { TxParams, Perm, TransferStatusCode, ErrorCode, FULL_DECIMALS } from '../../../../types';
+import {
+  TxParams,
+  Perm,
+  TransferStatusCode,
+  ErrorCode,
+  GetLogs,
+  Subscribe,
+  GetLogsAsyncParams,
+  SubscribeAsyncParams,
+  EventCallback,
+  FULL_DECIMALS,
+} from '../../../../types';
 import {
   numberToBigNumber,
   valueToWei,
@@ -37,6 +62,158 @@ export namespace VestingEscrowWalletTransactionParams {
   export interface AddScheduleFromTemplateMulti extends AddScheduleFromTemplateMultiParams {}
   export interface RevokeSchedulesMulti extends RevokeSchedulesMultiParams {}
   export interface ModifyScheduleMulti extends ModifyScheduleMultiParams {}
+}
+
+interface AddScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.AddSchedule;
+  callback: EventCallback<VestingEscrowWalletAddScheduleEventArgs_3_0_0>;
+}
+
+interface GetAddScheduleLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.AddSchedule;
+}
+
+interface ModifyScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.ModifySchedule;
+  callback: EventCallback<VestingEscrowWalletModifyScheduleEventArgs_3_0_0>;
+}
+
+interface GetModifyScheduleLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.ModifySchedule;
+}
+
+interface RevokeAllSchedulesSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RevokeAllSchedules;
+  callback: EventCallback<VestingEscrowWalletRevokeAllSchedulesEventArgs_3_0_0>;
+}
+
+interface GetRevokeAllSchedulesLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RevokeAllSchedules;
+}
+
+interface RevokeScheduleSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RevokeSchedule;
+  callback: EventCallback<VestingEscrowWalletRevokeScheduleEventArgs_3_0_0>;
+}
+
+interface GetRevokeScheduleLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RevokeSchedule;
+}
+
+interface DepositTokensSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.DepositTokens;
+  callback: EventCallback<VestingEscrowWalletDepositTokensEventArgs_3_0_0>;
+}
+
+interface GetDepositTokensLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.DepositTokens;
+}
+
+interface SendToTreasurySubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.SendToTreasury;
+  callback: EventCallback<VestingEscrowWalletSendToTreasuryEventArgs_3_0_0>;
+}
+
+interface GetSendToTreasuryLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.SendToTreasury;
+}
+
+interface SendTokensSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.SendTokens;
+  callback: EventCallback<VestingEscrowWalletSendTokensEventArgs_3_0_0>;
+}
+
+interface GetSendTokensLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.SendTokens;
+}
+
+interface AddTemplateSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.AddTemplate;
+  callback: EventCallback<VestingEscrowWalletAddTemplateEventArgs_3_0_0>;
+}
+
+interface GetAddTemplateLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.AddTemplate;
+}
+
+interface RemoveTemplateSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RemoveTemplate;
+  callback: EventCallback<VestingEscrowWalletRemoveTemplateEventArgs_3_0_0>;
+}
+
+interface GetRemoveTemplateLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.RemoveTemplate;
+}
+
+interface TreasuryWalletChangedSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.TreasuryWalletChanged;
+  callback: EventCallback<VestingEscrowWalletTreasuryWalletChangedEventArgs_3_0_0>;
+}
+
+interface GetTreasuryWalletChangedLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.TreasuryWalletChanged;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.Pause;
+  callback: EventCallback<VestingEscrowWalletPauseEventArgs_3_0_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.Unpause;
+  callback: EventCallback<VestingEscrowWalletUnpauseEventArgs_3_0_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: VestingEscrowWalletEvents_3_0_0.Unpause;
+}
+
+export interface VestingEscrowWalletSubscribeAsyncParams extends Subscribe {
+  (params: AddScheduleSubscribeAsyncParams): Promise<string>;
+  (params: ModifyScheduleSubscribeAsyncParams): Promise<string>;
+  (params: RevokeAllSchedulesSubscribeAsyncParams): Promise<string>;
+  (params: RevokeScheduleSubscribeAsyncParams): Promise<string>;
+  (params: DepositTokensSubscribeAsyncParams): Promise<string>;
+  (params: SendToTreasurySubscribeAsyncParams): Promise<string>;
+  (params: SendTokensSubscribeAsyncParams): Promise<string>;
+  (params: AddTemplateSubscribeAsyncParams): Promise<string>;
+  (params: RemoveTemplateSubscribeAsyncParams): Promise<string>;
+  (params: TreasuryWalletChangedSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+}
+
+export interface GetVestingEscrowWalletLogsAsyncParams extends GetLogs {
+  (params: GetAddScheduleLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletAddScheduleEventArgs_3_0_0>[]>;
+  (params: GetModifyScheduleLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletModifyScheduleEventArgs_3_0_0>[]
+  >;
+  (params: GetRevokeAllSchedulesLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletRevokeAllSchedulesEventArgs_3_0_0>[]
+  >;
+  (params: GetRevokeScheduleLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletRevokeScheduleEventArgs_3_0_0>[]
+  >;
+  (params: GetDepositTokensLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletDepositTokensEventArgs_3_0_0>[]
+  >;
+  (params: GetSendToTreasuryLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletSendToTreasuryEventArgs_3_0_0>[]
+  >;
+  (params: GetSendTokensLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletSendTokensEventArgs_3_0_0>[]>;
+  (params: GetAddTemplateLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletAddTemplateEventArgs_3_0_0>[]>;
+  (params: GetRemoveTemplateLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletRemoveTemplateEventArgs_3_0_0>[]
+  >;
+  (params: GetTreasuryWalletChangedLogsAsyncParams): Promise<
+    LogWithDecodedArgs<VestingEscrowWalletTreasuryWalletChangedEventArgs_3_0_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletPauseEventArgs_3_0_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<VestingEscrowWalletUnpauseEventArgs_3_0_0>[]>;
 }
 
 /**
@@ -843,5 +1020,4 @@ export default abstract class VestingEscrowWalletCommon extends ModuleCommon {
 
 export function isVestingEscrowWallet(wrapper: ContractWrapper): wrapper is VestingEscrowWalletCommon {
   return wrapper instanceof VestingEscrowWalletCommon;
-};
-
+}
