+++ src/index.ts
@@ -65,12 +65,52 @@ export {
   isModuleFactory,
   ModuleFactory_3_0_0,
   isModuleFactory_3_0_0,
+  ModuleFactoryEvents,
+  ModuleFactoryEventArgs,
+  ModuleFactoryGenerateModuleFromFactoryEventArgs,
+  ModuleFactoryOwnershipTransferredEventArgs,
+  ModuleFactoryChangeCostTypeEventArgs,
+  ModuleFactoryChangeSTVersionBoundEventArgs,
+  ModuleFactoryChangeSetupCostEventArgs,
 } from './contract_wrappers/modules/module_factory_wrapper';
 export {
   SecurityToken,
   isSecurityToken,
   SecurityToken_3_0_0,
   isSecurityToken_3_0_0,
+  SecurityTokenEvents,
+  SecurityTokenEventArgs,
+  SecurityTokenApprovalEventArgs,
+  SecurityTokenAuthorizedOperatorByPartitionEventArgs,
+  SecurityTokenAuthorizedOperatorEventArgs,
+  SecurityTokenCheckpointCreatedEventArgs,
+  SecurityTokenControllerRedemptionEventArgs,
+  SecurityTokenControllerTransferEventArgs,
+  SecurityTokenDisableControllerEventArgs,
+  SecurityTokenDocumentRemovedEventArgs,
+  SecurityTokenDocumentUpdatedEventArgs,
+  SecurityTokenFreezeIssuanceEventArgs,
+  SecurityTokenFreezeTransfersEventArgs,
+  SecurityTokenGranularityChangedEventArgs,
+  SecurityTokenIssuedByPartitionEventArgs,
+  SecurityTokenIssuedEventArgs,
+  SecurityTokenModuleAddedEventArgs,
+  SecurityTokenModuleArchivedEventArgs,
+  SecurityTokenModuleBudgetChangedEventArgs,
+  SecurityTokenModuleRemovedEventArgs,
+  SecurityTokenModuleUnarchivedEventArgs,
+  SecurityTokenOwnershipTransferredEventArgs,
+  SecurityTokenRedeemedByPartitionEventArgs,
+  SecurityTokenRedeemedEventArgs,
+  SecurityTokenRevokedOperatorByPartitionEventArgs,
+  SecurityTokenRevokedOperatorEventArgs,
+  SecurityTokenSetControllerEventArgs,
+  SecurityTokenTokenUpgradedEventArgs,
+  SecurityTokenTransferByPartitionEventArgs,
+  SecurityTokenTransferEventArgs,
+  SecurityTokenTreasuryWalletChangedEventArgs,
+  SecurityTokenUpdateTokenDetailsEventArgs,
+  SecurityTokenUpdateTokenNameEventArgs,
 } from './contract_wrappers/tokens/security_token_wrapper';
 export { default as ERC20 } from './contract_wrappers/tokens/erc20_detailed_wrapper';
 export {
@@ -78,12 +118,39 @@ export {
   isERC20DividendCheckpoint,
   ERC20DividendCheckpoint_3_0_0,
   isERC20DividendCheckpoint_3_0_0,
+  ERC20DividendCheckpointEvents,
+  ERC20DividendCheckpointEventArgs,
+  ERC20DividendCheckpointERC20DividendClaimedEventArgs,
+  ERC20DividendCheckpointERC20DividendDepositedEventArgs,
+  ERC20DividendCheckpointERC20DividendReclaimedEventArgs,
+  ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs,
+  ERC20DividendCheckpointPauseEventArgs,
+  ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs,
+  ERC20DividendCheckpointSetWalletEventArgs,
+  ERC20DividendCheckpointSetWithholdingEventArgs,
+  ERC20DividendCheckpointSetWithholdingFixedEventArgs,
+  ERC20DividendCheckpointUnpauseEventArgs,
+  ERC20DividendCheckpointUpdateDividendDatesEventArgs,
 } from './contract_wrappers/modules/checkpoint/erc20_dividend_checkpoint_wrapper';
 export {
   EtherDividendCheckpoint,
   isEtherDividendCheckpoint,
   EtherDividendCheckpoint_3_0_0,
   isEtherDividendCheckpoint_3_0_0,
+  EtherDividendCheckpointEvents,
+  EtherDividendCheckpointEventArgs,
+  EtherDividendCheckpointEtherDividendClaimFailedEventArgs,
+  EtherDividendCheckpointEtherDividendClaimedEventArgs,
+  EtherDividendCheckpointEtherDividendDepositedEventArgs,
+  EtherDividendCheckpointEtherDividendReclaimedEventArgs,
+  EtherDividendCheckpointEtherDividendWithholdingWithdrawnEventArgs,
+  EtherDividendCheckpointPauseEventArgs,
+  EtherDividendCheckpointSetDefaultExcludedAddressesEventArgs,
+  EtherDividendCheckpointSetWalletEventArgs,
+  EtherDividendCheckpointSetWithholdingEventArgs,
+  EtherDividendCheckpointSetWithholdingFixedEventArgs,
+  EtherDividendCheckpointUnpauseEventArgs,
+  EtherDividendCheckpointUpdateDividendDatesEventArgs,
 } from './contract_wrappers/modules/checkpoint/ether_dividend_checkpoint_wrapper';
 export {
   GeneralPermissionManager,
@@ -92,6 +159,12 @@ export {
   isGeneralPermissionManager_3_0_0,
   GeneralPermissionManager_3_1_0,
   isGeneralPermissionManager_3_1_0,
+  GeneralPermissionManagerEvents,
+  GeneralPermissionManagerEventArgs,
+  GeneralPermissionManagerAddDelegateEventArgs,
+  GeneralPermissionManagerChangePermissionEventArgs,
+  GeneralPermissionManagerPauseEventArgs,
+  GeneralPermissionManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/permission_manager/general_permission_manager_wrapper';
 export {
   CappedSTO,
@@ -100,14 +173,43 @@ export {
   isCappedSTO_3_0_0,
   CappedSTO_3_1_0,
   isCappedSTO_3_1_0,
+  CappedSTOEvents,
+  CappedSTOEventArgs,
+  CappedSTOAllowPreMintFlagEventArgs,
+  CappedSTOPauseEventArgs,
+  CappedSTOReserveTokenMintEventArgs,
+  CappedSTOReserveTokenTransferEventArgs,
+  CappedSTORevokePreMintFlagEventArgs,
+  CappedSTOSetAllowBeneficialInvestmentsEventArgs,
+  CappedSTOTokenPurchaseEventArgs,
+  CappedSTOSetFundRaiseTypesEventArgs,
+  CappedSTOUnpauseEventArgs,
 } from './contract_wrappers/modules/sto/capped_sto_wrapper';
-export {  
+export {
   USDTieredSTO,
   isUSDTieredSTO,
   USDTieredSTO_3_0_0,
   isUSDTieredSTO_3_0_0,
   USDTieredSTO_3_1_0,
   isUSDTieredSTO_3_1_0,
+  USDTieredSTOEvents,
+  USDTieredSTOEventArgs,
+  USDTieredSTOAllowPreMintFlagEventArgs,
+  USDTieredSTOPauseEventArgs,
+  USDTieredSTOFundsReceivedEventArgs,
+  USDTieredSTOReserveTokenMintEventArgs,
+  USDTieredSTOReserveTokenTransferEventArgs,
+  USDTieredSTORevokePreMintFlagEventArgs,
+  USDTieredSTOSetAddressesEventArgs,
+  USDTieredSTOSetAllowBeneficialInvestmentsEventArgs,
+  USDTieredSTOSetFundRaiseTypesEventArgs,
+  USDTieredSTOSetLimitsEventArgs,
+  USDTieredSTOSetNonAccreditedLimitEventArgs,
+  USDTieredSTOSetTiersEventArgs,
+  USDTieredSTOSetTimesEventArgs,
+  USDTieredSTOSetTreasuryWalletEventArgs,
+  USDTieredSTOTokenPurchaseEventArgs,
+  USDTieredSTOUnpauseEventArgs,
 } from './contract_wrappers/modules/sto/usd_tiered_sto_wrapper';
 export {
   GeneralTransferManager,
@@ -116,48 +218,116 @@ export {
   isGeneralTransferManager_3_0_0,
   GeneralTransferManager_3_1_0,
   isGeneralTransferManager_3_1_0,
+  GeneralTransferManagerEvents,
+  GeneralTransferManagerEventArgs,
+  GeneralTransferManagerChangeDefaultsEventArgs,
+  GeneralTransferManagerChangeIssuanceAddressEventArgs,
+  GeneralTransferManagerModifyInvestorFlagEventArgs,
+  GeneralTransferManagerModifyKYCDataEventArgs,
+  GeneralTransferManagerModifyTransferRequirementsEventArgs,
+  GeneralTransferManagerPauseEventArgs,
+  GeneralTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/general_transfer_manager_wrapper';
 export {
   CountTransferManager,
   isCountTransferManager,
   CountTransferManager_3_0_0,
   isCountTransferManager_3_0_0,
+  CountTransferManagerEvents,
+  CountTransferManagerEventArgs,
+  CountTransferManagerModifyHolderCountEventArgs,
+  CountTransferManagerPauseEventArgs,
+  CountTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper';
 export {
   ManualApprovalTransferManager,
   isManualApprovalTransferManager,
   ManualApprovalTransferManager_3_0_0,
   isManualApprovalTransferManager_3_0_0,
+  ManualApprovalTransferManagerEvents,
+  ManualApprovalTransferManagerEventArgs,
+  ManualApprovalTransferManagerAddManualApprovalEventArgs,
+  ManualApprovalTransferManagerModifyManualApprovalEventArgs,
+  ManualApprovalTransferManagerPauseEventArgs,
+  ManualApprovalTransferManagerRevokeManualApprovalEventArgs,
+  ManualApprovalTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/manual_approval_transfer_manager_wrapper';
 export {
   PercentageTransferManager,
   isPercentageTransferManager,
   PercentageTransferManager_3_0_0,
   isPercentageTransferManager_3_0_0,
+  PercentageTransferManagerEvents,
+  PercentageTransferManagerEventArgs,
+  PercentageTransferManagerModifyHolderPercentageEventArgs,
+  PercentageTransferManagerModifyWhitelistEventArgs,
+  PercentageTransferManagerPauseEventArgs,
+  PercentageTransferManagerSetAllowPrimaryIssuanceEventArgs,
+  PercentageTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/percentage_transfer_manager_wrapper';
 export {
   LockUpTransferManager,
   isLockUpTransferManager,
   LockUpTransferManager_3_0_0,
   isLockUpTransferManager_3_0_0,
+  LockUpTransferManagerEvents,
+  LockUpTansferManagerEventArgs,
+  LockUpTransferManagerAddLockUpToUserEventArgs,
+  LockUpTransferManagerAddNewLockUpTypeEventArgs,
+  LockUpTransferManagerModifyLockUpTypeEventArgs,
+  LockUpTransferManagerPauseEventArgs,
+  LockUpTransferManagerRemoveLockUpFromUserEventArgs,
+  LockUpTransferManagerRemoveLockUpTypeEventArgs,
+  LockUpTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/lock_up_transfer_manager_wrapper';
 export {
   BlacklistTransferManager,
   isBlacklistTransferManager,
   BlacklistTransferManager_3_0_0,
   isBlacklistTransferManager_3_0_0,
+  BlacklistTransferManagerEvents,
+  BlacklistTransferManagerEventArgs,
+  BlacklistTransferManagerAddBlacklistTypeEventArgs,
+  BlacklistTransferManagerAddInvestorToBlacklistEventArgs,
+  BlacklistTransferManagerDeleteBlacklistTypeEventArgs,
+  BlacklistTransferManagerDeleteInvestorFromBlacklistEventArgs,
+  BlacklistTransferManagerModifyBlacklistTypeEventArgs,
+  BlacklistTransferManagerPauseEventArgs,
+  BlacklistTransferManagerUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/blacklist_transfer_manager_wrapper';
 export {
   VolumeRestrictionTransferManager,
   isVolumeRestrictionTransferManager,
   VolumeRestrictionTransferManager_3_0_0,
   isVolumeRestrictionTransferManager_3_0_0,
+  VolumeRestrictionTMEvents,
+  VolumeRestrictionTMEventArgs,
+  VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs,
+  VolumeRestrictionTMAddDefaultRestrictionEventArgs,
+  VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs,
+  VolumeRestrictionTMAddIndividualRestrictionEventArgs,
+  VolumeRestrictionTMChangedExemptWalletListEventArgs,
+  VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs,
+  VolumeRestrictionTMDefaultRestrictionRemovedEventArgs,
+  VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs,
+  VolumeRestrictionTMIndividualRestrictionRemovedEventArgs,
+  VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs,
+  VolumeRestrictionTMModifyDefaultRestrictionEventArgs,
+  VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs,
+  VolumeRestrictionTMModifyIndividualRestrictionEventArgs,
+  VolumeRestrictionTMPauseEventArgs,
+  VolumeRestrictionTMUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/volume_restriction_transfer_manager_wrapper';
 export {
   RestrictedPartialSaleTransferManager,
   isRestrictedPartialSaleTransferManager,
   RestrictedPartialSaleTransferManager_3_1_0,
   isRestrictedPartialSaleTransferManager_3_1_0,
+  RestrictedPartialSaleTransferManagerEvents,
+  RestrictedPartialSaleTransferManagerEventArgs,
+  RestrictedPartialSaleTMChangedExemptWalletListEventArgs,
+  RestrictedPartialSaleTMPauseEventArgs,
+  RestrictedPartialSaleTMUnpauseEventArgs,
 } from './contract_wrappers/modules/transfer_manager/restricted_partial_sale_transfer_manager_wrapper';
 export {
   VestingEscrowWallet,
@@ -166,6 +336,20 @@ export {
   isVestingEscrowWallet_3_0_0,
   VestingEscrowWallet_3_1_0,
   isVestingEscrowWallet_3_1_0,
+  VestingEscrowWalletEvents,
+  VestingEscrowWalletEventArgs,
+  VestingEscrowWalletAddScheduleEventArgs,
+  VestingEscrowWalletAddTemplateEventArgs,
+  VestingEscrowWalletDepositTokensEventArgs,
+  VestingEscrowWalletModifyScheduleEventArgs,
+  VestingEscrowWalletPauseEventArgs,
+  VestingEscrowWalletRemoveTemplateEventArgs,
+  VestingEscrowWalletRevokeAllSchedulesEventArgs,
+  VestingEscrowWalletRevokeScheduleEventArgs,
+  VestingEscrowWalletSendToTreasuryEventArgs,
+  VestingEscrowWalletSendTokensEventArgs,
+  VestingEscrowWalletTreasuryWalletChangedEventArgs,
+  VestingEscrowWalletUnpauseEventArgs,
 } from './contract_wrappers/modules/wallet/vesting_escrow_wallet_wrapper';
 export * from './types';
 export { PolymathAPI } from './PolymathAPI';
@@ -211,269 +395,23 @@ export {
   FeatureRegistryChangeFeatureStatusEventArgs_3_0_0 as FeatureRegistryChangeFeatureStatusEventArgs,
   FeatureRegistryOwnershipTransferredEventArgs_3_0_0 as FeatureRegistryOwnershipTransferredEventArgs,
   ISecurityTokenContract_3_0_0 as SecurityTokenContract_3_0_0,
-  ISecurityTokenEvents_3_0_0 as SecurityTokenEvents_3_0_0,
-  ISecurityTokenEventArgs_3_0_0 as SecurityTokenEventArgs_3_0_0,
-  ISecurityTokenApprovalEventArgs_3_0_0 as SecurityTokenApprovalEventArgs_3_0_0,
-  ISecurityTokenAuthorizedOperatorByPartitionEventArgs_3_0_0 as SecurityTokenAuthorizedOperatorByPartitionEventArgs_3_0_0,
-  ISecurityTokenAuthorizedOperatorEventArgs_3_0_0 as SecurityTokenAuthorizedOperatorEventArgs_3_0_0,
-  ISecurityTokenCheckpointCreatedEventArgs_3_0_0 as SecurityTokenCheckpointCreatedEventArgs_3_0_0,
-  ISecurityTokenControllerRedemptionEventArgs_3_0_0 as SecurityTokenControllerRedemptionEventArgs_3_0_0,
-  ISecurityTokenControllerTransferEventArgs_3_0_0 as SecurityTokenControllerTransferEventArgs_3_0_0,
-  ISecurityTokenDisableControllerEventArgs_3_0_0 as SecurityTokenDisableControllerEventArgs_3_0_0,
-  ISecurityTokenDocumentRemovedEventArgs_3_0_0 as SecurityTokenDocumentRemovedEventArgs_3_0_0,
-  ISecurityTokenDocumentUpdatedEventArgs_3_0_0 as SecurityTokenDocumentUpdatedEventArgs_3_0_0,
-  ISecurityTokenFreezeIssuanceEventArgs_3_0_0 as SecurityTokenFreezeIssuanceEventArgs_3_0_0,
-  ISecurityTokenFreezeTransfersEventArgs_3_0_0 as SecurityTokenFreezeTransfersEventArgs_3_0_0,
-  ISecurityTokenGranularityChangedEventArgs_3_0_0 as SecurityTokenGranularityChangedEventArgs_3_0_0,
-  ISecurityTokenIssuedByPartitionEventArgs_3_0_0 as SecurityTokenIssuedByPartitionEventArgs_3_0_0,
-  ISecurityTokenIssuedEventArgs_3_0_0 as SecurityTokenIssuedEventArgs_3_0_0,
-  ISecurityTokenModuleAddedEventArgs_3_0_0 as SecurityTokenModuleAddedEventArgs_3_0_0,
-  ISecurityTokenModuleArchivedEventArgs_3_0_0 as SecurityTokenModuleArchivedEventArgs_3_0_0,
-  ISecurityTokenModuleBudgetChangedEventArgs_3_0_0 as SecurityTokenModuleBudgetChangedEventArgs_3_0_0,
-  ISecurityTokenModuleRemovedEventArgs_3_0_0 as SecurityTokenModuleRemovedEventArgs_3_0_0,
-  ISecurityTokenModuleUnarchivedEventArgs_3_0_0 as SecurityTokenModuleUnarchivedEventArgs_3_0_0,
-  ISecurityTokenOwnershipTransferredEventArgs_3_0_0 as SecurityTokenOwnershipTransferredEventArgs_3_0_0,
-  ISecurityTokenRedeemedByPartitionEventArgs_3_0_0 as SecurityTokenRedeemedByPartitionEventArgs_3_0_0,
-  ISecurityTokenRedeemedEventArgs_3_0_0 as SecurityTokenRedeemedEventArgs_3_0_0,
-  ISecurityTokenRevokedOperatorByPartitionEventArgs_3_0_0 as SecurityTokenRevokedOperatorByPartitionEventArgs_3_0_0,
-  ISecurityTokenRevokedOperatorEventArgs_3_0_0 as SecurityTokenRevokedOperatorEventArgs_3_0_0,
-  ISecurityTokenSetControllerEventArgs_3_0_0 as SecurityTokenSetControllerEventArgs_3_0_0,
-  ISecurityTokenTokenUpgradedEventArgs_3_0_0 as SecurityTokenTokenUpgradedEventArgs_3_0_0,
-  ISecurityTokenTransferByPartitionEventArgs_3_0_0 as SecurityTokenTransferByPartitionEventArgs_3_0_0,
-  ISecurityTokenTransferEventArgs_3_0_0 as SecurityTokenTransferEventArgs_3_0_0,
-  ISecurityTokenTreasuryWalletChangedEventArgs_3_0_0 as SecurityTokenTreasuryWalletChangedEventArgs_3_0_0,
-  ISecurityTokenUpdateTokenDetailsEventArgs_3_0_0 as SecurityTokenUpdateTokenDetailsEventArgs_3_0_0,
-  ISecurityTokenUpdateTokenNameEventArgs_3_0_0 as SecurityTokenUpdateTokenNameEventArgs_3_0_0,  
-  ISecurityTokenRegistryContract_3_0_0 as SecurityTokenRegistryContract,
-  ISecurityTokenRegistryEvents_3_0_0 as SecurityTokenRegistryEvents,
-  ISecurityTokenRegistryEventArgs_3_0_0 as SecurityTokenRegistryEventArgs,
-  ISecurityTokenRegistryChangeExpiryLimitEventArgs_3_0_0 as SecurityTokenRegistryChangeExpiryLimitEventArgs,
-  ISecurityTokenRegistryChangeSecurityLaunchFeeEventArgs_3_0_0 as SecurityTokenRegistryChangeSecurityLaunchFeeEventArgs,
-  ISecurityTokenRegistryChangeTickerRegistrationFeeEventArgs_3_0_0 as SecurityTokenRegistryChangeTickerRegistrationFeeEventArgs,
-  ISecurityTokenRegistryNewSecurityTokenEventArgs_3_0_0 as SecurityTokenRegistryNewSecurityTokenEventArgs,
-  ISecurityTokenRegistryOwnershipTransferredEventArgs_3_0_0 as SecurityTokenRegistryOwnershipTransferredEventArgs,
-  ISecurityTokenRegistryPauseEventArgs_3_0_0 as SecurityTokenRegistryPauseEventArgs,
-  ISecurityTokenRegistryUnpauseEventArgs_3_0_0 as SecurityTokenRegistryUnpauseEventArgs,
-  ISecurityTokenRegistryRegisterTickerEventArgs_3_0_0 as SecurityTokenRegistryRegisterTickerEventArgs,
-  ISecurityTokenRegistryTickerRemovedEventArgs_3_0_0 as SecurityTokenRegistryTickerRemovedEventArgs,
-  ISecurityTokenRegistryChangeTickerOwnershipEventArgs_3_0_0 as SecurityTokenRegistryChangeTickerOwnershipEventArgs,
-  ISecurityTokenRegistryChangeFeeCurrencyEventArgs_3_0_0 as SecurityTokenRegistryChangeFeeCurrencyEventArgs,
-  ISecurityTokenRegistrySecurityTokenRefreshedEventArgs_3_0_0 as SecurityTokenRegistrySecurityTokenRefreshedEventArgs,
-  ISecurityTokenRegistryProtocolFactorySetEventArgs_3_0_0 as SecurityTokenRegistryProtocolFactorySetEventArgs,
-  ISecurityTokenRegistryLatestVersionSetEventArgs_3_0_0 as SecurityTokenRegistryLatestVersionSetEventArgs,
-  ISecurityTokenRegistryProtocolFactoryRemovedEventArgs_3_0_0 as SecurityTokenRegistryProtocolFactoryRemovedEventArgs,
   GeneralTransferManagerContract_3_0_0,
-  GeneralTransferManagerEventArgs_3_0_0,
-  GeneralTransferManagerEvents_3_0_0,
-  GeneralTransferManagerChangeIssuanceAddressEventArgs_3_0_0,
-  GeneralTransferManagerChangeDefaultsEventArgs_3_0_0,
-  GeneralTransferManagerPauseEventArgs_3_0_0,
-  GeneralTransferManagerUnpauseEventArgs_3_0_0,
-  GeneralTransferManagerModifyKYCDataEventArgs_3_0_0,
-  GeneralTransferManagerModifyInvestorFlagEventArgs_3_0_0,
-  GeneralTransferManagerModifyTransferRequirementsEventArgs_3_0_0,
-  GeneralPermissionManagerContract_3_0_0,
-  GeneralPermissionManagerEventArgs_3_0_0,
-  GeneralPermissionManagerEvents_3_0_0,
-  GeneralPermissionManagerChangePermissionEventArgs_3_0_0,
-  GeneralPermissionManagerAddDelegateEventArgs_3_0_0,
-  GeneralPermissionManagerPauseEventArgs_3_0_0,
-  GeneralPermissionManagerUnpauseEventArgs_3_0_0,
-  ERC20DividendCheckpointContract_3_0_0,
-  ERC20DividendCheckpointEventArgs_3_0_0,
-  ERC20DividendCheckpointEvents_3_0_0,
-  ERC20DividendCheckpointERC20DividendDepositedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendClaimedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendReclaimedEventArgs_3_0_0,
-  ERC20DividendCheckpointERC20DividendWithholdingWithdrawnEventArgs_3_0_0,
-  ERC20DividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0,
-  ERC20DividendCheckpointSetWithholdingEventArgs_3_0_0,
-  ERC20DividendCheckpointSetWithholdingFixedEventArgs_3_0_0,
-  ERC20DividendCheckpointPauseEventArgs_3_0_0,
-  ERC20DividendCheckpointSetWalletEventArgs_3_0_0,
-  ERC20DividendCheckpointUnpauseEventArgs_3_0_0,
-  ERC20DividendCheckpointUpdateDividendDatesEventArgs_3_0_0,
-  CappedSTOContract_3_0_0,
-  CappedSTOEventArgs_3_0_0,
-  CappedSTOEvents_3_0_0,
-  CappedSTOPauseEventArgs_3_0_0,
-  CappedSTOUnpauseEventArgs_3_0_0,
-  CappedSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0,
-  CappedSTOTokenPurchaseEventArgs_3_0_0,
-  CappedSTOSetFundRaiseTypesEventArgs_3_0_0,
-  USDTieredSTOContract_3_0_0,
-  USDTieredSTOEventArgs_3_0_0,
-  USDTieredSTOEvents_3_0_0,
-  USDTieredSTOFundsReceivedEventArgs_3_0_0,
-  USDTieredSTOPauseEventArgs_3_0_0,
-  USDTieredSTOReserveTokenMintEventArgs_3_0_0,
-  USDTieredSTOSetAddressesEventArgs_3_0_0,
-  USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_0_0,
-  USDTieredSTOSetFundRaiseTypesEventArgs_3_0_0,
-  USDTieredSTOSetLimitsEventArgs_3_0_0,
-  USDTieredSTOSetNonAccreditedLimitEventArgs_3_0_0,
-  USDTieredSTOSetTiersEventArgs_3_0_0,
-  USDTieredSTOSetTimesEventArgs_3_0_0,
-  USDTieredSTOSetTreasuryWalletEventArgs_3_0_0,
-  USDTieredSTOTokenPurchaseEventArgs_3_0_0,
-  USDTieredSTOUnpauseEventArgs_3_0_0,
-  ManualApprovalTransferManagerContract_3_0_0,
-  ManualApprovalTransferManagerEventArgs_3_0_0,
-  ManualApprovalTransferManagerEvents_3_0_0,
-  ManualApprovalTransferManagerAddManualApprovalEventArgs_3_0_0,
-  ManualApprovalTransferManagerModifyManualApprovalEventArgs_3_0_0,
-  ManualApprovalTransferManagerRevokeManualApprovalEventArgs_3_0_0,
-  ManualApprovalTransferManagerPauseEventArgs_3_0_0,
-  ManualApprovalTransferManagerUnpauseEventArgs_3_0_0,
-  PercentageTransferManagerContract_3_0_0,
-  PercentageTransferManagerEventArgs_3_0_0,
-  PercentageTransferManagerEvents_3_0_0,
-  PercentageTransferManagerModifyHolderPercentageEventArgs_3_0_0,
-  PercentageTransferManagerModifyWhitelistEventArgs_3_0_0,
-  PercentageTransferManagerPauseEventArgs_3_0_0,
-  PercentageTransferManagerSetAllowPrimaryIssuanceEventArgs_3_0_0,
-  PercentageTransferManagerUnpauseEventArgs_3_0_0,
-  LockUpTransferManagerContract_3_0_0,
-  LockUpTransferManagerEventArgs_3_0_0,
-  LockUpTransferManagerEvents_3_0_0,
-  LockUpTransferManagerAddLockUpToUserEventArgs_3_0_0,
-  LockUpTransferManagerRemoveLockUpFromUserEventArgs_3_0_0,
-  LockUpTransferManagerModifyLockUpTypeEventArgs_3_0_0,
-  LockUpTransferManagerAddNewLockUpTypeEventArgs_3_0_0,
-  LockUpTransferManagerRemoveLockUpTypeEventArgs_3_0_0,
-  LockUpTransferManagerPauseEventArgs_3_0_0,
-  LockUpTransferManagerUnpauseEventArgs_3_0_0,
-  BlacklistTransferManagerContract_3_0_0,
-  BlacklistTransferManagerEventArgs_3_0_0,
-  BlacklistTransferManagerEvents_3_0_0,
-  BlacklistTransferManagerAddBlacklistTypeEventArgs_3_0_0,
-  BlacklistTransferManagerModifyBlacklistTypeEventArgs_3_0_0,
-  BlacklistTransferManagerDeleteBlacklistTypeEventArgs_3_0_0,
-  BlacklistTransferManagerAddInvestorToBlacklistEventArgs_3_0_0,
-  BlacklistTransferManagerDeleteInvestorFromBlacklistEventArgs_3_0_0,
-  BlacklistTransferManagerPauseEventArgs_3_0_0,
-  BlacklistTransferManagerUnpauseEventArgs_3_0_0,
-  CountTransferManagerContract_3_0_0,
-  CountTransferManagerEventArgs_3_0_0,
-  CountTransferManagerEvents_3_0_0,
-  CountTransferManagerModifyHolderCountEventArgs_3_0_0,
-  CountTransferManagerPauseEventArgs_3_0_0,
-  CountTransferManagerUnpauseEventArgs_3_0_0,
-  EtherDividendCheckpointContract_3_0_0,
-  EtherDividendCheckpointEventArgs_3_0_0,
-  EtherDividendCheckpointEvents_3_0_0,
-  EtherDividendCheckpointEtherDividendDepositedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendClaimedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendReclaimedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendClaimFailedEventArgs_3_0_0,
-  EtherDividendCheckpointEtherDividendWithholdingWithdrawnEventArgs_3_0_0,
-  EtherDividendCheckpointSetDefaultExcludedAddressesEventArgs_3_0_0,
-  EtherDividendCheckpointSetWithholdingEventArgs_3_0_0,
-  EtherDividendCheckpointSetWithholdingFixedEventArgs_3_0_0,
-  EtherDividendCheckpointSetWalletEventArgs_3_0_0,
-  EtherDividendCheckpointUpdateDividendDatesEventArgs_3_0_0,
-  EtherDividendCheckpointPauseEventArgs_3_0_0,
-  EtherDividendCheckpointUnpauseEventArgs_3_0_0,
-  VolumeRestrictionTMContract_3_0_0,
-  VolumeRestrictionTMEventArgs_3_0_0,
-  VolumeRestrictionTMEvents_3_0_0,
-  VolumeRestrictionTMAddDefaultDailyRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMAddDefaultRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMAddIndividualDailyRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMAddIndividualRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMChangedExemptWalletListEventArgs_3_0_0,
-  VolumeRestrictionTMDefaultDailyRestrictionRemovedEventArgs_3_0_0,
-  VolumeRestrictionTMDefaultRestrictionRemovedEventArgs_3_0_0,
-  VolumeRestrictionTMIndividualDailyRestrictionRemovedEventArgs_3_0_0,
-  VolumeRestrictionTMIndividualRestrictionRemovedEventArgs_3_0_0,
-  VolumeRestrictionTMModifyDefaultDailyRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMModifyDefaultRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMModifyIndividualDailyRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMModifyIndividualRestrictionEventArgs_3_0_0,
-  VolumeRestrictionTMPauseEventArgs_3_0_0,
-  VolumeRestrictionTMUnpauseEventArgs_3_0_0,
-  RestrictedPartialSaleTMPauseEventArgs_3_1_0,
-  RestrictedPartialSaleTMUnpauseEventArgs_3_1_0,
-  RestrictedPartialSaleTMContract_3_1_0,
-  RestrictedPartialSaleTMEventArgs_3_1_0,
-  RestrictedPartialSaleTMEvents_3_1_0,
-  RestrictedPartialSaleTMChangedExemptWalletListEventArgs_3_1_0,
-  CappedSTOContract_3_1_0,
-  CappedSTOEventArgs_3_1_0,
-  CappedSTOEvents_3_1_0,
-  CappedSTOPauseEventArgs_3_1_0,
-  CappedSTOUnpauseEventArgs_3_1_0,
-  CappedSTOSetAllowBeneficialInvestmentsEventArgs_3_1_0,
-  CappedSTOTokenPurchaseEventArgs_3_1_0,
-  CappedSTOSetFundRaiseTypesEventArgs_3_1_0,
-  CappedSTOAllowPreMintFlagEventArgs_3_1_0,
-  CappedSTOReserveTokenMintEventArgs_3_1_0,
-  CappedSTORevokePreMintFlagEventArgs_3_1_0,
-  CappedSTOReserveTokenTransferEventArgs_3_1_0,
-  USDTieredSTOContract_3_1_0,
-  USDTieredSTOEventArgs_3_1_0,
-  USDTieredSTOEvents_3_1_0,
-  USDTieredSTOFundsReceivedEventArgs_3_1_0,
-  USDTieredSTOPauseEventArgs_3_1_0,
-  USDTieredSTOReserveTokenMintEventArgs_3_1_0,
-  USDTieredSTOSetAddressesEventArgs_3_1_0,
-  USDTieredSTOSetAllowBeneficialInvestmentsEventArgs_3_1_0,
-  USDTieredSTOSetFundRaiseTypesEventArgs_3_1_0,
-  USDTieredSTOSetLimitsEventArgs_3_1_0,
-  USDTieredSTOSetNonAccreditedLimitEventArgs_3_1_0,
-  USDTieredSTOSetTiersEventArgs_3_1_0,
-  USDTieredSTOSetTimesEventArgs_3_1_0,
-  USDTieredSTOSetTreasuryWalletEventArgs_3_1_0,
-  USDTieredSTOTokenPurchaseEventArgs_3_1_0,
-  USDTieredSTOUnpauseEventArgs_3_1_0,
-  USDTieredSTOAllowPreMintFlagEventArgs_3_1_0,
-  USDTieredSTORevokePreMintFlagEventArgs_3_1_0,
-  USDTieredSTOReserveTokenTransferEventArgs_3_1_0,
   GeneralTransferManagerContract_3_1_0,
-  GeneralTransferManagerEventArgs_3_1_0,
-  GeneralTransferManagerEvents_3_1_0,
-  GeneralTransferManagerChangeIssuanceAddressEventArgs_3_1_0,
-  GeneralTransferManagerChangeDefaultsEventArgs_3_1_0,
-  GeneralTransferManagerPauseEventArgs_3_1_0,
-  GeneralTransferManagerUnpauseEventArgs_3_1_0,
-  GeneralTransferManagerModifyKYCDataEventArgs_3_1_0,
-  GeneralTransferManagerModifyInvestorFlagEventArgs_3_1_0,
-  GeneralTransferManagerModifyTransferRequirementsEventArgs_3_1_0,
+  GeneralPermissionManagerContract_3_0_0,
   GeneralPermissionManagerContract_3_1_0,
-  GeneralPermissionManagerEventArgs_3_1_0,
-  GeneralPermissionManagerEvents_3_1_0,
-  GeneralPermissionManagerChangePermissionEventArgs_3_1_0,
-  GeneralPermissionManagerAddDelegateEventArgs_3_1_0,
-  GeneralPermissionManagerPauseEventArgs_3_1_0,
-  GeneralPermissionManagerUnpauseEventArgs_3_1_0,
+  ERC20DividendCheckpointContract_3_0_0,
+  EtherDividendCheckpointContract_3_0_0,
+  CappedSTOContract_3_0_0,
+  CappedSTOContract_3_1_0,
+  USDTieredSTOContract_3_0_0,
+  USDTieredSTOContract_3_1_0,
+  ManualApprovalTransferManagerContract_3_0_0,
+  PercentageTransferManagerContract_3_0_0,
+  LockUpTransferManagerContract_3_0_0,
+  BlacklistTransferManagerContract_3_0_0,
+  CountTransferManagerContract_3_0_0,
+  VolumeRestrictionTMContract_3_0_0,
+  RestrictedPartialSaleTMContract_3_1_0,
   VestingEscrowWalletContract_3_0_0,
-  VestingEscrowWalletEventArgs_3_0_0,
-  VestingEscrowWalletEvents_3_0_0,
-  VestingEscrowWalletPauseEventArgs_3_0_0,
-  VestingEscrowWalletUnpauseEventArgs_3_0_0,
-  VestingEscrowWalletAddScheduleEventArgs_3_0_0,
-  VestingEscrowWalletModifyScheduleEventArgs_3_0_0,
-  VestingEscrowWalletRevokeAllSchedulesEventArgs_3_0_0,
-  VestingEscrowWalletRevokeScheduleEventArgs_3_0_0,
-  VestingEscrowWalletDepositTokensEventArgs_3_0_0,
-  VestingEscrowWalletSendToTreasuryEventArgs_3_0_0,
-  VestingEscrowWalletSendTokensEventArgs_3_0_0,
-  VestingEscrowWalletAddTemplateEventArgs_3_0_0,
-  VestingEscrowWalletRemoveTemplateEventArgs_3_0_0,
-  VestingEscrowWalletTreasuryWalletChangedEventArgs_3_0_0,
   VestingEscrowWalletContract_3_1_0,
-  VestingEscrowWalletEventArgs_3_1_0,
-  VestingEscrowWalletEvents_3_1_0,
-  VestingEscrowWalletPauseEventArgs_3_1_0,
-  VestingEscrowWalletUnpauseEventArgs_3_1_0,
-  VestingEscrowWalletAddScheduleEventArgs_3_1_0,
-  VestingEscrowWalletModifyScheduleEventArgs_3_1_0,
-  VestingEscrowWalletRevokeAllSchedulesEventArgs_3_1_0,
-  VestingEscrowWalletRevokeScheduleEventArgs_3_1_0,
-  VestingEscrowWalletDepositTokensEventArgs_3_1_0,
-  VestingEscrowWalletSendToTreasuryEventArgs_3_1_0,
-  VestingEscrowWalletSendTokensEventArgs_3_1_0,
-  VestingEscrowWalletAddTemplateEventArgs_3_1_0,
-  VestingEscrowWalletRemoveTemplateEventArgs_3_1_0,
-  VestingEscrowWalletTreasuryWalletChangedEventArgs_3_1_0,
 } from '@polymathnetwork/abi-wrappers';
